Add more try/catches to the script editor specifically
