+++ src/SerialLoops/Editors/ScriptEditor.cs
@@ -886,1138 +886,1061 @@ namespace SerialLoops.Editors
 
         private void CommandsPanel_SelectedItemChanged(object sender, EventArgs e)
         {
-            ScriptItemCommand command = ((ScriptCommandSectionEntry)((ScriptCommandSectionTreeGridView)sender).SelectedItem).Command;
-            _editorControls.Items.Clear();
-
-            _addCommandButton.Enabled = true;
-            _addSectionButton.Enabled = true;
-            _deleteButton.Enabled = true;
-
-            // if we've selected a script section header
-            if (command is null)
+            try
             {
-                return;
-            }
+                ScriptItemCommand command = ((ScriptCommandSectionEntry)((ScriptCommandSectionTreeGridView)sender).SelectedItem).Command;
+                _editorControls.Items.Clear();
 
-            Application.Instance.Invoke(() => UpdatePreview());
+                _addCommandButton.Enabled = true;
+                _addSectionButton.Enabled = true;
+                _deleteButton.Enabled = true;
 
-            if (command.Parameters.Count == 0)
-            {
-                return;
-            }
-
-            int cols = 1;
-
-            // We're going to embed table layouts inside a table layout so we can have colspan
-            TableLayout controlsTable = new()
-            {
-                Spacing = new Size(5, 5)
-            };
-            controlsTable.Rows.Add(new(new TableLayout { Spacing = new Size(5, 5) }));
-            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows.Add(new());
-
-            int currentRow = 0, currentCol = 0, currentShort = 0;
-            for (int i = 0; i < command.Parameters.Count; i++)
-            {
-                ScriptParameter parameter = command.Parameters[i];
-                switch (parameter.Type)
+                // if we've selected a script section header
+                if (command is null)
                 {
-                    case ScriptParameter.ParameterType.BG:
-                        BgScriptParameter bgParam = (BgScriptParameter)parameter;
-                        CommandGraphicSelectionButton bgSelectionButton = new(bgParam.Background is not null ? bgParam.Background
-                            : NonePreviewableGraphic.BACKGROUND, _tabs, _log)
-                        {
-                            Command = command,
-                            ParameterIndex = i,
-                            Project = _project,
-                        };
-
-                        // BGDISPTEMP is able to display a lot more kinds of backgrounds properly than the other BG commands
-                        // Hence, this switch to make sure you don't accidentally crash the game
-                        if (command.Verb == CommandVerb.BG_FADE)
-                        {
-                            bgSelectionButton.Items.Add(NonePreviewableGraphic.BACKGROUND);
-                            switch (i)
-                            {
-                                case 0:
-                                    bgSelectionButton.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Background &&
-                                        (((BackgroundItem)i).BackgroundType == BgType.TEX_BG))
-                                        .Select(b => b as IPreviewableGraphic));
-                                    break;
-                                case 1:
-                                    bgSelectionButton.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Background &&
-                                        ((BackgroundItem)i).BackgroundType != BgType.KINETIC_SCREEN && ((BackgroundItem)i).BackgroundType != BgType.TEX_BG)
-                                        .Select(b => b as IPreviewableGraphic));
-                                    break;
-                            }
-                        }
-                        else
-                        {
-                            switch (command.Verb)
-                            {
-                                case CommandVerb.BG_DISP:
-                                case CommandVerb.BG_DISP2:
-                                    bgSelectionButton.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Background &&
-                                        (((BackgroundItem)i).BackgroundType == BgType.TEX_BG))
-                                        .Select(b => b as IPreviewableGraphic));
-                                    break;
-
-                                case CommandVerb.BG_DISPCG:
-                                    bgSelectionButton.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Background &&
-                                        ((BackgroundItem)i).BackgroundType != BgType.KINETIC_SCREEN && ((BackgroundItem)i).BackgroundType != BgType.TEX_BG)
-                                        .Select(b => b as IPreviewableGraphic));
-                                    break;
-
-                                case CommandVerb.KBG_DISP:
-                                    bgSelectionButton.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Background &&
-                                        ((BackgroundItem)i).BackgroundType == BgType.KINETIC_SCREEN).Select(b => b as IPreviewableGraphic));
-                                    break;
-                            }
-                        }
-                        bgSelectionButton.SelectedChanged.Executed += (obj, args) => BgSelectionButton_SelectionMade(bgSelectionButton, args);
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, bgSelectionButton));
-                        break;
-
-                    case ScriptParameter.ParameterType.BG_SCROLL_DIRECTION:
-                        ScriptCommandDropDown bgScrollDropDown = new() { Command = command, ParameterIndex = i };
-                        bgScrollDropDown.Items.AddRange(Enum.GetNames<BgScrollDirectionScriptParameter.BgScrollDirection>().Select(i => new ListItem { Text = i, Key = i }));
-                        bgScrollDropDown.SelectedKey = ((BgScrollDirectionScriptParameter)parameter).ScrollDirection.ToString();
-                        bgScrollDropDown.SelectedKeyChanged += BgScrollDropDown_SelectedKeyChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, bgScrollDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.BGM:
-                        BgmScriptParameter bgmParam = (BgmScriptParameter)parameter;
-                        StackLayout bgmLink = ControlGenerator.GetFileLink(bgmParam.Bgm, _tabs, _log);
-
-                        ScriptCommandDropDown bgmDropDown = new() { Command = command, ParameterIndex = i, Link = (ClearableLinkButton)bgmLink.Items[1].Control };
-                        bgmDropDown.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.BGM).Select(i => new ListItem { Text = i.DisplayName, Key = i.DisplayName }));
-                        bgmDropDown.SelectedKey = bgmParam.Bgm.DisplayName;
-                        bgmDropDown.SelectedKeyChanged += BgmDropDown_SelectedKeyChanged;
-
-                        StackLayout bgmLayout = new()
-                        {
-                            Orientation = Orientation.Horizontal,
-                            VerticalContentAlignment = VerticalAlignment.Center,
-                            Items =
-                            {
-                                ControlGenerator.GetControlWithLabel(parameter.Name, bgmDropDown),
-                                bgmLink,
-                            }
-                        };
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(bgmLayout);
-                        break;
-
-                    case ScriptParameter.ParameterType.BGM_MODE:
-                        ScriptCommandDropDown bgmModeDropDown = new() { Command = command, ParameterIndex = i };
-                        bgmModeDropDown.Items.AddRange(Enum.GetNames<BgmModeScriptParameter.BgmMode>().Select(i => new ListItem { Text = i, Key = i }));
-                        bgmModeDropDown.SelectedKey = ((BgmModeScriptParameter)parameter).Mode.ToString();
-                        bgmModeDropDown.SelectedKeyChanged += BgmModeDropDown_SelectedKeyChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, bgmModeDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.BOOL:
-                        ScriptCommandCheckBox boolParameterCheckbox = new() { Command = command, ParameterIndex = i, Checked = ((BoolScriptParameter)parameter).Value };
-                        boolParameterCheckbox.CheckedChanged += BoolParameterCheckbox_CheckedChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, boolParameterCheckbox));
-                        break;
-
-                    case ScriptParameter.ParameterType.CHARACTER:
-                        ScriptCommandDropDown dialoguePropertyDropDown = new() { Command = command, ParameterIndex = i };
-                        dialoguePropertyDropDown.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Character)
-                            .Select(c => new ListItem { Key = c.DisplayName, Text = c.DisplayName[4..] }));
-                        dialoguePropertyDropDown.SelectedKey = ((DialoguePropertyScriptParameter)parameter).Character.Name;
-                        dialoguePropertyDropDown.SelectedKeyChanged += DialoguePropertyDropDown_SelectedKeyChanged;
-                        _currentSpeakerDropDown.OtherDropDowns.Add(dialoguePropertyDropDown);
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, dialoguePropertyDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.CHESS_FILE:
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name,
-                            new TextBox { Text = ((ChessFileScriptParameter)parameter).ChessFileIndex.ToString() }));
-                        break;
-
-                    case ScriptParameter.ParameterType.CHESS_PIECE:
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name,
-                            new TextBox { Text = ((ChessPieceScriptParameter)parameter).ChessPiece.ToString() }));
-                        break;
-
-                    case ScriptParameter.ParameterType.CHESS_SPACE:
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name,
-                            new TextBox { Text = ((ChessSpaceScriptParameter)parameter).SpaceIndex.ToString() }));
-                        break;
-
-                    case ScriptParameter.ParameterType.CHIBI:
-                        ScriptCommandDropDown chibiDropDown = new() { Command = command, ParameterIndex = i };
-                        chibiDropDown.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Chibi).Select(i => new ListItem { Text = i.Name, Key = i.Name }));
-                        chibiDropDown.SelectedKey = ((ChibiScriptParameter)parameter).Chibi.Name;
-                        chibiDropDown.SelectedKeyChanged += ChibiDropDown_SelectedKeyChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, chibiDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.CHIBI_EMOTE:
-                        ScriptCommandDropDown chibiEmoteDropDown = new() { Command = command, ParameterIndex = i };
-                        chibiEmoteDropDown.Items.AddRange(Enum.GetNames<ChibiEmoteScriptParameter.ChibiEmote>().Select(i => new ListItem { Text = i, Key = i }));
-                        chibiEmoteDropDown.SelectedKey = ((ChibiEmoteScriptParameter)parameter).Emote.ToString();
-                        chibiEmoteDropDown.SelectedKeyChanged += ChibiEmoteDropDown_SelectedKeyChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, chibiEmoteDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.CHIBI_ENTER_EXIT:
-                        ScriptCommandDropDown chibiEnterExitDropDown = new() { Command = command, ParameterIndex = i };
-                        chibiEnterExitDropDown.Items.AddRange(Enum.GetNames<ChibiEnterExitScriptParameter.ChibiEnterExitType>().Select(i => new ListItem { Text = i, Key = i }));
-                        chibiEnterExitDropDown.SelectedKey = ((ChibiEnterExitScriptParameter)parameter).Mode.ToString();
-                        chibiEnterExitDropDown.SelectedKeyChanged += ChibiEnterExitDropDown_SelectedKeyChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, chibiEnterExitDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.COLOR:
-                        ScriptCommandColorPicker colorPicker = new()
-                        {
-                            AllowAlpha = false,
-                            Value = ((ColorScriptParameter)parameter).Color.ToEtoDrawingColor(),
-                            Command = command,
-                            ParameterIndex = i,
-                        };
-                        colorPicker.ValueChanged += ColorPicker_ValueChanged;
-                        currentShort += 2;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, colorPicker));
-                        break;
-
-                    case ScriptParameter.ParameterType.CONDITIONAL:
-                        ScriptCommandTextBox conditionalBox = new() { Text = ((ConditionalScriptParameter)parameter).Conditional, Command = command, ParameterIndex = i };
-                        conditionalBox.TextChanged += ConditionalBox_TextChanged;
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name,
-                            conditionalBox));
-                        break;
-
-                    case ScriptParameter.ParameterType.COLOR_MONOCHROME:
-                        ScriptCommandDropDown colorMonochromeDropDown = new() { Command = command, ParameterIndex = i, CurrentShort = currentShort };
-                        colorMonochromeDropDown.Items.AddRange(Enum.GetNames<ColorMonochromeScriptParameter.ColorMonochrome>().Select(i => new ListItem { Text = i, Key = i }));
-                        colorMonochromeDropDown.SelectedKey = ((ColorMonochromeScriptParameter)parameter).ColorType.ToString();
-                        colorMonochromeDropDown.SelectedKeyChanged += ColorMonochromeDropDown_SelectedKeyChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, colorMonochromeDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.DIALOGUE:
-                        DialogueScriptParameter dialogueParam = (DialogueScriptParameter)parameter;
-                        ScriptCommandDropDown speakerDropDown = new() { Command = command, ParameterIndex = i, OtherDropDowns = new() };
-                        speakerDropDown.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Character).Select(c => new ListItem { Key = c.DisplayName, Text = c.DisplayName[4..] }));
-                        try
-                        {
-                            speakerDropDown.SelectedKey = _project.Items.First(i => i.Type == ItemDescription.ItemType.Character && i.DisplayName == $"CHR_{_project.Characters[(int)dialogueParam.Line.Speaker].Name}").DisplayName;
-                        }
-                        catch (InvalidOperationException)
-                        {
-                            _log.LogError("Failed to find character item -- have you saved all of your changes to character names?");
-                            return;
-                        }
-
-                        speakerDropDown.SelectedKeyChanged += SpeakerDropDown_SelectedKeyChanged;
-                        if (currentCol > 0)
-                        {
-                            controlsTable.Rows.Add(new(new TableLayout { Spacing = new Size(5, 5) }));
-                            currentRow++;
-                            currentCol = 0;
-                        }
-                        _currentSpeakerDropDown = speakerDropDown;
-
-                        ScriptCommandTextArea dialogueTextArea = new()
-                        {
-                            Text = dialogueParam.Line.Text.GetSubstitutedString(_project),
-                            AcceptsReturn = true,
-                            Command = command,
-                            ParameterIndex = i,
-                        };
-                        dialogueTextArea.TextChanged += DialogueTextArea_TextChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabelTable(parameter.Name,
-                            new StackLayout
-                            {
-                                Orientation = Orientation.Horizontal,
-                                Items =
-                                {
-                                    speakerDropDown,
-                                    new StackLayoutItem(dialogueTextArea, expand: true),
-                                },
-                            }));
-                        controlsTable.Rows.Add(new(new TableLayout { Spacing = new Size(5, 5) }));
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows.Add(new());
-                        currentRow++;
-                        currentCol = 0;
-                        break;
-
-                    case ScriptParameter.ParameterType.EPISODE_HEADER:
-                        ScriptCommandDropDown epHeaderDropDown = new() { Command = command, ParameterIndex = i };
-                        epHeaderDropDown.Items.AddRange(Enum.GetNames<EpisodeHeaderScriptParameter.Episode>().Select(e => new ListItem { Key = e, Text = e }));
-                        epHeaderDropDown.SelectedKey = ((EpisodeHeaderScriptParameter)parameter).EpisodeHeaderIndex.ToString();
-                        epHeaderDropDown.SelectedKeyChanged += EpHeaderDropDown_SelectedKeyChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name,
-                            epHeaderDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.FLAG:
-                        ScriptCommandTextBox flagTextBox = new() { Command = command, ParameterIndex = i, Text = ((FlagScriptParameter)parameter).FlagName };
-                        flagTextBox.TextChanged += FlagTextBox_TextChanged;
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, flagTextBox));
-                        break;
-
-                    case ScriptParameter.ParameterType.FRIENDSHIP_LEVEL:
-                        ScriptCommandDropDown friendshipLevelDropDown = new() { Command = command, ParameterIndex = i };
-                        friendshipLevelDropDown.Items.AddRange(Enum.GetNames<FriendshipLevelScriptParameter.FriendshipCharacter>().Select(f => new ListItem { Key = f, Text = f }));
-                        friendshipLevelDropDown.SelectedKey = ((FriendshipLevelScriptParameter)parameter).Character.ToString();
-                        friendshipLevelDropDown.SelectedIndexChanged += FriendshipLevelDropDown_SelectedIndexChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, friendshipLevelDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.ITEM:
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name,
-                            new TextBox { Text = ((ItemScriptParameter)parameter).ItemIndex.ToString() }));
-                        break;
-
-                    case ScriptParameter.ParameterType.MAP:
-                        MapScriptParameter mapParam = (MapScriptParameter)parameter;
-                        ScriptCommandDropDown mapDropDown = new() { Command = command, ParameterIndex = i };
-                        mapDropDown.Items.Add(new ListItem { Text = "NONE", Key = "NONE" });
-                        mapDropDown.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Map).Select(i => new ListItem { Text = i.Name, Key = i.Name }));
-                        mapDropDown.SelectedKey = mapParam.Map?.Name ?? "NONE";
-                        mapDropDown.SelectedKeyChanged += MapDropDown_SelectedKeyChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, mapDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.OPTION:
-                        OptionScriptParameter optionParam = (OptionScriptParameter)parameter;
-                        ScriptCommandDropDown optionDropDown = new() { Command = command, ParameterIndex = i };
-                        optionDropDown.Items.Add(new ListItem { Text = "NONE", Key = "0" });
-                        optionDropDown.Items.AddRange(_script.Event.ChoicesSection.Objects.Skip(1).SkipLast(1).Select(c => new ListItem { Text = c.Text.GetSubstitutedString(_project), Key = c.Id.ToString() }));
-                        optionDropDown.SelectedKey = optionParam.Option.Id.ToString();
-                        optionDropDown.SelectedKeyChanged += OptionDropDown_SelectedKeyChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, optionDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.PALETTE_EFFECT:
-                        ScriptCommandDropDown paletteEffectDropDown = new() { Command = command, ParameterIndex = i };
-                        paletteEffectDropDown.Items.AddRange(Enum.GetNames<PaletteEffectScriptParameter.PaletteEffect>().Select(t => new ListItem { Text = t, Key = t }));
-                        paletteEffectDropDown.SelectedKey = ((PaletteEffectScriptParameter)parameter).Effect.ToString();
-                        paletteEffectDropDown.SelectedKeyChanged += PaletteEffectDropDown_SelectedKeyChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, paletteEffectDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.PLACE:
-                        PlaceScriptParameter placeParam = (PlaceScriptParameter)parameter;
-                        CommandGraphicSelectionButton placeSelectionButton = new(placeParam.Place is not null ? placeParam.Place
-                            : NonePreviewableGraphic.PLACE, _tabs, _log)
-                        {
-                            Command = command,
-                            ParameterIndex = i,
-                            Project = _project,
-                        };
-                        placeSelectionButton.Items.Add(NonePreviewableGraphic.PLACE);
-                        placeSelectionButton.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Place)
-                            .Select(p => p as IPreviewableGraphic));
-                        placeSelectionButton.SelectedChanged.Executed += (obj, args) => PlaceSelectionButtonSelectedChanged_Executed(placeSelectionButton, args);
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, placeSelectionButton));
-                        break;
-
-                    case ScriptParameter.ParameterType.SCREEN:
-                        ScriptCommandScreenSelector screenSelector = new(_log, ((ScreenScriptParameter)parameter).Screen, true)
-                        {
-                            Command = command,
-                            ParameterIndex = i,
-                            CurrentShort = currentShort,
-                        };
-                        screenSelector.ScreenChanged += ScreenSelector_ScreenChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, screenSelector));
-                        break;
-
-                    case ScriptParameter.ParameterType.SCRIPT_SECTION:
-                        ScriptCommandDropDown scriptSectionDropDown = new() { Command = command, ParameterIndex = i, CurrentShort = i };
-                        if (command.Verb == CommandVerb.VGOTO)
-                        {
-                            scriptSectionDropDown.CurrentShort = 2;
-                        }
-                        scriptSectionDropDown.Items.AddRange(_script.Event.ScriptSections.Where(s => (_script.Event.LabelsSection.Objects.FirstOrDefault(l => l.Name.Replace("/", "") == s.Name)?.Id ?? 0) != 0).Select(s => new ListItem { Text = s.Name, Key = s.Name }));
-                        scriptSectionDropDown.SelectedKey = ((ScriptSectionScriptParameter)parameter)?.Section?.Name ?? "NONE";
-                        scriptSectionDropDown.SelectedKeyChanged += ScriptSectionDropDown_SelectedKeyChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, scriptSectionDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.SFX:
-                        ScriptCommandNumericStepper sfxNumericStepper = new() { Command = command, ParameterIndex = i, Value = ((SfxScriptParameter)parameter).SfxIndex };
-                        sfxNumericStepper.ValueChanged += SfxNumericStepper_ValueChanged;
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, sfxNumericStepper));
-                        break;
-
-                    case ScriptParameter.ParameterType.SFX_MODE:
-                        ScriptCommandDropDown sfxModeDropDown = new() { Command = command, ParameterIndex = i };
-                        sfxModeDropDown.Items.AddRange(Enum.GetNames<SfxModeScriptParameter.SfxMode>().Select(t => new ListItem { Text = t, Key = t }));
-                        sfxModeDropDown.SelectedKey = ((SfxModeScriptParameter)parameter).Mode.ToString();
-                        sfxModeDropDown.SelectedKeyChanged += SfxModeDropDown_SelectedKeyChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, sfxModeDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.SHORT:
-                        ScriptCommandNumericStepper shortNumericStepper = new()
-                        {
-                            Command = command,
-                            ParameterIndex = i,
-                            MinValue = short.MinValue,
-                            MaxValue = short.MaxValue,
-                            DecimalPlaces = 0,
-                            Value = ((ShortScriptParameter)parameter).Value
-                        };
-                        if (command.Verb == CommandVerb.SND_PLAY && parameter.Name == "Crossfade Time (Frames)")
-                        {
-                            shortNumericStepper.SecondIndex = 4;
-                        }
-                        if (command.Verb == CommandVerb.HARUHI_METER)
-                        {
-                            shortNumericStepper.ParameterIndex++;
-                        }
-                        shortNumericStepper.ValueChanged += ShortNumericStepper_ValueChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, shortNumericStepper));
-                        break;
-
-                    case ScriptParameter.ParameterType.SPRITE:
-                        SpriteScriptParameter spriteParam = (SpriteScriptParameter)parameter;
-                        CommandGraphicSelectionButton spriteSelectionButton = new(spriteParam.Sprite is not null ? spriteParam.Sprite
-                            : NonePreviewableGraphic.CHARACTER_SPRITE, _tabs, _log)
-                        {
-                            Command = command,
-                            ParameterIndex = i,
-                            Project = _project,
-                        };
-                        spriteSelectionButton.Items.Add(NonePreviewableGraphic.CHARACTER_SPRITE);
-                        spriteSelectionButton.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Character_Sprite).Select(s => (IPreviewableGraphic)s));
-                        spriteSelectionButton.SelectedChanged.Executed += (obj, args) => SpriteSelectionButton_SelectionMade(spriteSelectionButton, args);
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, spriteSelectionButton));
-                        break;
-
-                    case ScriptParameter.ParameterType.SPRITE_ENTRANCE:
-                        ScriptCommandDropDown spriteEntranceDropDown = new() { Command = command, ParameterIndex = i };
-                        spriteEntranceDropDown.Items.AddRange(Enum.GetNames<SpriteEntranceScriptParameter.SpriteEntranceTransition>().Select(t => new ListItem { Text = t, Key = t }));
-                        spriteEntranceDropDown.SelectedKey = ((SpriteEntranceScriptParameter)parameter).EntranceTransition.ToString();
-                        spriteEntranceDropDown.SelectedKeyChanged += SpriteEntranceDropDown_SelectedKeyChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, spriteEntranceDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.SPRITE_EXIT:
-                        ScriptCommandDropDown spriteExitDropDown = new() { Command = command, ParameterIndex = i };
-                        spriteExitDropDown.Items.AddRange(Enum.GetNames<SpriteExitScriptParameter.SpriteExitTransition>().Select(t => new ListItem { Text = t, Key = t }));
-                        spriteExitDropDown.SelectedKey = ((SpriteExitScriptParameter)parameter).ExitTransition.ToString();
-                        spriteExitDropDown.SelectedKeyChanged += SpriteExitDropDown_SelectedKeyChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, spriteExitDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.SPRITE_SHAKE:
-                        ScriptCommandDropDown spriteShakeDropDown = new() { Command = command, ParameterIndex = i };
-                        spriteShakeDropDown.Items.AddRange(Enum.GetNames<SpriteShakeScriptParameter.SpriteShakeEffect>().Select(t => new ListItem { Text = t, Key = t }));
-                        spriteShakeDropDown.SelectedKey = ((SpriteShakeScriptParameter)parameter).ShakeEffect.ToString();
-                        spriteShakeDropDown.SelectedKeyChanged += SpriteShakeDropDown_SelectedKeyChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, spriteShakeDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.TEXT_ENTRANCE_EFFECT:
-                        ScriptCommandDropDown textEntranceEffectDropDown = new() { Command = command, ParameterIndex = i };
-                        textEntranceEffectDropDown.Items.AddRange(Enum.GetNames<TextEntranceEffectScriptParameter.TextEntranceEffect>().Select(t => new ListItem { Text = t, Key = t }));
-                        textEntranceEffectDropDown.SelectedKey = ((TextEntranceEffectScriptParameter)parameter).EntranceEffect.ToString();
-                        textEntranceEffectDropDown.SelectedKeyChanged += TextEntranceEffectDropDown_SelectedKeyChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, textEntranceEffectDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.TOPIC:
-                        string topicName = _project.Items.FirstOrDefault(i => i.Type == ItemDescription.ItemType.Topic &&
-                            ((TopicItem)i).Topic.Id == ((TopicScriptParameter)parameter).TopicId)?.DisplayName;
-                        if (string.IsNullOrEmpty(topicName))
-                        {
-                            // If the topic has been deleted, we will just display the index in a textbox
-                            TopicSelectButton setUpTopicControlButton = new() { Text = "Select a Topic", ScriptCommand = command, ParameterIndex = i };
-
-                            StackLayout deletedTopicLayout = new()
-                            {
-                                Orientation = Orientation.Horizontal,
-                                Items =
-                                {
-                                    new TextBox { Text = ((TopicScriptParameter)parameter).TopicId.ToString() },
-                                    setUpTopicControlButton,
-                                },
-                            };
-
-                            setUpTopicControlButton.Layout = deletedTopicLayout;
-                            setUpTopicControlButton.Click += SetUpTopicControlButton_Click;
-
-                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                                ControlGenerator.GetControlWithLabel(parameter.Name,
-                                deletedTopicLayout));
-                        }
-                        else
-                        {
-                            StackLayout topicLink = ControlGenerator.GetFileLink(_project.Items.FirstOrDefault(i => i.Type == ItemDescription.ItemType.Topic &&
-                                ((TopicItem)i).Topic.Id == ((TopicScriptParameter)parameter).TopicId), _tabs, _log);
-
-                            ScriptCommandDropDown topicDropDown = new() { Command = command, ParameterIndex = i, Link = (ClearableLinkButton)topicLink.Items[1].Control };
-                            topicDropDown.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Topic)
-                                .Select(t => new ListItem { Key = t.DisplayName, Text = t.DisplayName }));
-                            topicDropDown.SelectedKey = topicName;
-                            topicDropDown.SelectedIndexChanged += TopicDropDown_SelectedIndexChanged;
-
-                            StackLayout topicLinkLayout = new()
-                            {
-                                Orientation = Orientation.Horizontal,
-                                Items =
-                                {
-                                    topicDropDown,
-                                    topicLink,
-                                },
-                            };
-
-                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                                ControlGenerator.GetControlWithLabel(parameter.Name,
-                                topicLinkLayout));
-                        }
-                        break;
-
-                    case ScriptParameter.ParameterType.TRANSITION:
-                        ScriptCommandDropDown transitionDropDown = new() { Command = command, ParameterIndex = i };
-                        transitionDropDown.Items.AddRange(Enum.GetNames<TransitionScriptParameter.TransitionEffect>().Select(t => new ListItem { Text = t, Key = t }));
-                        transitionDropDown.SelectedKey = ((TransitionScriptParameter)parameter).Transition.ToString();
-                        transitionDropDown.SelectedKeyChanged += TransitionDropDown_SelectedKeyChanged;
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
-                            ControlGenerator.GetControlWithLabel(parameter.Name, transitionDropDown));
-                        break;
-
-                    case ScriptParameter.ParameterType.VOICE_LINE:
-                        VoicedLineScriptParameter vceParam = (VoicedLineScriptParameter)parameter;
-                        StackLayout vceLink = ControlGenerator.GetFileLink(vceParam.VoiceLine is not null ? vceParam.VoiceLine : NoneItem.VOICE, _tabs, _log);
-
-                        ScriptCommandDropDown vceDropDown = new() { Command = command, ParameterIndex = i, Link = (ClearableLinkButton)vceLink.Items[1].Control };
-                        vceDropDown.Items.Add(new ListItem { Key = "NONE", Text = "NONE" });
-                        vceDropDown.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Voice).Select(i => new ListItem { Text = i.Name, Key = i.Name }));
-                        vceDropDown.SelectedKey = vceParam.VoiceLine?.Name ?? "NONE";
-                        vceDropDown.SelectedKeyChanged += VceDropDown_SelectedKeyChanged;
-
-                        StackLayout vceLayout = new()
-                        {
-                            Orientation = Orientation.Horizontal,
-                            VerticalContentAlignment = VerticalAlignment.Center,
-                            Items =
-                            {
-                                ControlGenerator.GetControlWithLabel(parameter.Name, vceDropDown),
-                                vceLink,
-                            }
-                        };
-
-                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(vceLayout);
-                        break;
-
-                    default:
-                        _log.LogError($"Invalid parameter detected in script {_script.Name} parameter {parameter.Name}");
-                        break;
-                }
-                currentCol++;
-                if (currentCol >= cols)
-                {
-                    currentCol = 0;
-                    currentRow++;
-                    controlsTable.Rows.Add(new(new TableLayout { Spacing = new Size(5, 5) }));
-                    ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows.Add(new());
-                }
-                currentShort++;
-            }
-
-            _editorControls.Items.Add(new StackLayoutItem(controlsTable, expand: true));
-        }
-
-        private void UpdatePreview()
-        {
-            _preview.Items.Clear();
-
-            SKBitmap previewBitmap = new(256, 384);
-            SKCanvas canvas = new(previewBitmap);
-            canvas.DrawColor(SKColors.Black);
-
-            ScriptItemCommand currentCommand = ((ScriptCommandSectionEntry)_commandsPanel.Viewer.SelectedItem)?.Command;
-            if (currentCommand is not null)
-            {
-                List<ScriptItemCommand> commands = currentCommand.WalkCommandGraph(_commands, _script.Graph);
-
-                if (commands is null)
-                {
-                    _log.LogWarning($"No path found to current command.");
-                    using Stream noPreviewStream = Assembly.GetCallingAssembly().GetManifestResourceStream("SerialLoops.Graphics.ScriptPreviewError.png");
-                    canvas.DrawImage(SKImage.FromEncodedData(noPreviewStream), new SKPoint(0, 0));
-                    canvas.Flush();
-                    _preview.Items.Add(new SKGuiImage(previewBitmap));
                     return;
                 }
 
-                if (commands.Any(c => c.Verb == CommandVerb.EPHEADER)
-                    && ((EpisodeHeaderScriptParameter)commands.Last(c => c.Verb == CommandVerb.EPHEADER).Parameters[0])
-                        .EpisodeHeaderIndex != EpisodeHeaderScriptParameter.Episode.None)
-                {
-                    // We use names here because display names can change but names cannot
-                    SystemTextureItem systexItem = ((EpisodeHeaderScriptParameter)commands.Last(c => c.Verb == CommandVerb.EPHEADER).Parameters[0]).EpisodeHeaderIndex switch
-                    {
-                        EpisodeHeaderScriptParameter.Episode.EPISODE_1 => 
-                            (SystemTextureItem)_project.Items.First(i => i.Name == "SYSTEX_SYS_CMN_T60"),
-                        EpisodeHeaderScriptParameter.Episode.EPISODE_2 =>
-                            (SystemTextureItem)_project.Items.First(i => i.Name == "SYSTEX_SYS_CMN_T61"),
-                        EpisodeHeaderScriptParameter.Episode.EPISODE_3 =>
-                            (SystemTextureItem)_project.Items.First(i => i.Name == "SYSTEX_SYS_CMN_T62"),
-                        EpisodeHeaderScriptParameter.Episode.EPISODE_4 =>
-                            (SystemTextureItem)_project.Items.First(i => i.Name == "SYSTEX_SYS_CMN_T63"),
-                        EpisodeHeaderScriptParameter.Episode.EPISODE_5 =>
-                            (SystemTextureItem)_project.Items.First(i => i.Name == "SYSTEX_SYS_CMN_T64"),
-                        EpisodeHeaderScriptParameter.Episode.EPILOGUE =>
-                            (SystemTextureItem)_project.Items.First(i => i.Name == "SYSTEX_SYS_CMN_T66"),
-                        _ => null,
-                    };
+                Application.Instance.Invoke(() => UpdatePreview());
 
-                    canvas.DrawBitmap(systexItem.GetTexture(), new SKPoint(0, 0));
+                if (command.Parameters.Count == 0)
+                {
+                    return;
                 }
-                else
+
+                int cols = 1;
+
+                // We're going to embed table layouts inside a table layout so we can have colspan
+                TableLayout controlsTable = new()
                 {
-                    // Draw top screen "kinetic" background
-                    for (int i = commands.Count - 1; i >= 0; i--)
-                    {
-                        if (commands[i].Verb == CommandVerb.KBG_DISP && ((BgScriptParameter)commands[i].Parameters[0]).Background is not null)
-                        {
-                            canvas.DrawBitmap(((BgScriptParameter)commands[i].Parameters[0]).Background.GetBackground(), new SKPoint(0, 0));
-                            break;
-                        }
-                    }
+                    Spacing = new Size(5, 5)
+                };
+                controlsTable.Rows.Add(new(new TableLayout { Spacing = new Size(5, 5) }));
+                ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows.Add(new());
 
-                    // Draw Place
-                    for (int i = commands.Count - 1; i >= 0; i--)
+                int currentRow = 0, currentCol = 0, currentShort = 0;
+                for (int i = 0; i < command.Parameters.Count; i++)
+                {
+                    ScriptParameter parameter = command.Parameters[i];
+                    switch (parameter.Type)
                     {
-                        if (commands[i].Verb == CommandVerb.SET_PLACE)
-                        {
-                            if (((BoolScriptParameter)commands[i].Parameters[0]).Value && (((PlaceScriptParameter)commands[i].Parameters[1]).Place is not null))
+                        case ScriptParameter.ParameterType.BG:
+                            BgScriptParameter bgParam = (BgScriptParameter)parameter;
+                            CommandGraphicSelectionButton bgSelectionButton = new(bgParam.Background is not null ? bgParam.Background
+                                : NonePreviewableGraphic.BACKGROUND, _tabs, _log)
                             {
-                                canvas.DrawBitmap(((PlaceScriptParameter)commands[i].Parameters[1]).Place.GetPreview(_project), new SKPoint(5, 40));
-                            }
-                            break;
-                        }
-                    }
+                                Command = command,
+                                ParameterIndex = i,
+                                Project = _project,
+                            };
 
-                    // Draw top screen chibis
-                    List<ChibiItem> chibis = new();
-
-                    foreach (StartingChibiEntry chibi in _script.Event.StartingChibisSection?.Objects ?? new List<StartingChibiEntry>())
-                    {
-                        if (chibi.ChibiIndex > 0)
-                        {
-                            chibis.Add((ChibiItem)_project.Items.First(i => i.Type == ItemDescription.ItemType.Chibi && ((ChibiItem)i).ChibiIndex == chibi.ChibiIndex));
-                        }
-                    }
-                    for (int i = 0; i < commands.Count; i++)
-                    {
-                        if (commands[i].Verb == CommandVerb.CHIBI_ENTEREXIT)
-                        {
-                            if (((ChibiEnterExitScriptParameter)commands[i].Parameters[1]).Mode == ChibiEnterExitScriptParameter.ChibiEnterExitType.ENTER)
+                            // BGDISPTEMP is able to display a lot more kinds of backgrounds properly than the other BG commands
+                            // Hence, this switch to make sure you don't accidentally crash the game
+                            if (command.Verb == CommandVerb.BG_FADE)
                             {
-                                ChibiItem chibi = ((ChibiScriptParameter)commands[i].Parameters[0]).Chibi;
-                                if (!chibis.Contains(chibi))
+                                bgSelectionButton.Items.Add(NonePreviewableGraphic.BACKGROUND);
+                                switch (i)
                                 {
-                                    if (chibi.ChibiIndex < 1 || chibis.Count == 0)
-                                    {
-                                        chibis.Add(chibi);
-                                    }
-                                    else
-                                    {
-                                        bool inserted = false;
-                                        for (int j = 0; j < chibis.Count; j++)
-                                        {
-                                            if (chibis[j].ChibiIndex > chibi.ChibiIndex)
-                                            {
-                                                chibis.Insert(j, chibi);
-                                                inserted = true;
-                                                break;
-                                            }
-                                        }
-                                        if (!inserted)
-                                        {
-                                            chibis.Add(chibi);
-                                        }
-                                    }
-                                }
-                                else
-                                {
-                                    _log.LogWarning($"Chibi {chibi.Name} set to join, but already was present");
+                                    case 0:
+                                        bgSelectionButton.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Background &&
+                                            (((BackgroundItem)i).BackgroundType == BgType.TEX_BG))
+                                            .Select(b => b as IPreviewableGraphic));
+                                        break;
+                                    case 1:
+                                        bgSelectionButton.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Background &&
+                                            ((BackgroundItem)i).BackgroundType != BgType.KINETIC_SCREEN && ((BackgroundItem)i).BackgroundType != BgType.TEX_BG)
+                                            .Select(b => b as IPreviewableGraphic));
+                                        break;
                                 }
                             }
                             else
                             {
-                                try
+                                switch (command.Verb)
                                 {
-                                    chibis.Remove(((ChibiScriptParameter)commands[i].Parameters[0]).Chibi);
+                                    case CommandVerb.BG_DISP:
+                                    case CommandVerb.BG_DISP2:
+                                        bgSelectionButton.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Background &&
+                                            (((BackgroundItem)i).BackgroundType == BgType.TEX_BG))
+                                            .Select(b => b as IPreviewableGraphic));
+                                        break;
+
+                                    case CommandVerb.BG_DISPCG:
+                                        bgSelectionButton.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Background &&
+                                            ((BackgroundItem)i).BackgroundType != BgType.KINETIC_SCREEN && ((BackgroundItem)i).BackgroundType != BgType.TEX_BG)
+                                            .Select(b => b as IPreviewableGraphic));
+                                        break;
+
+                                    case CommandVerb.KBG_DISP:
+                                        bgSelectionButton.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Background &&
+                                            ((BackgroundItem)i).BackgroundType == BgType.KINETIC_SCREEN).Select(b => b as IPreviewableGraphic));
+                                        break;
                                 }
-                                catch (Exception)
-                                {
-                                    _log.LogWarning($"Chibi set to leave was not present.");
-                                }
                             }
-                        }
-                    }
+                            bgSelectionButton.SelectedChanged.Executed += (obj, args) => BgSelectionButton_SelectionMade(bgSelectionButton, args);
 
-                    int chibiStartX, chibiY;
-                    if (commands.Any(c => c.Verb == EventFile.CommandVerb.OP_MODE))
-                    {
-                        chibiStartX = 100;
-                        chibiY = 50;
-                    }
-                    else
-                    {
-                        chibiStartX = 44;
-                        chibiY = 100;
-                    }
-                    int chibiCurrentX = chibiStartX;
-                    int chibiWidth = 0;
-                    foreach (ChibiItem chibi in chibis)
-                    {
-                        SKBitmap chibiFrame = chibi.ChibiAnimations.First().Value.ElementAt(0).Frame;
-                        canvas.DrawBitmap(chibiFrame, new SKPoint(chibiCurrentX, chibiY));
-                        chibiWidth = chibiFrame.Width - 10;
-                        chibiCurrentX += chibiWidth;
-                    }
-
-                    // Draw top screen chibi emotes
-                    if (currentCommand.Verb == CommandVerb.CHIBI_EMOTE)
-                    {
-                        ChibiItem chibi = ((ChibiScriptParameter)currentCommand.Parameters[0]).Chibi;
-                        if (chibis.Contains(chibi))
-                        {
-                            int chibiIndex = chibis.IndexOf(chibi);
-                            SKBitmap emotes = _project.Grp.Files.First(f => f.Name == "SYS_ADV_T08DNX").GetImage(width: 32, transparentIndex: 0);
-                            int internalYOffset = ((int)((ChibiEmoteScriptParameter)currentCommand.Parameters[1]).Emote - 1) * 32;
-                            int externalXOffset = chibiStartX + chibiWidth * chibiIndex;
-                            canvas.DrawBitmap(emotes, new SKRect(0, internalYOffset, 32, internalYOffset + 32), new SKRect(externalXOffset + 16, chibiY - 32, externalXOffset + 48, chibiY));
-                        }
-                        else
-                        {
-                            _log.LogWarning($"Chibi {chibi.Name} not currently on screen; cannot display emote.");
-                        }
-                    }
-                }
-
-                // Draw background
-                bool bgReverted = false;
-                ScriptItemCommand palCommand = commands.LastOrDefault(c => c.Verb == CommandVerb.PALEFFECT);
-                ScriptItemCommand lastBgCommand = commands.LastOrDefault(c => c.Verb == CommandVerb.BG_DISP ||
-                    c.Verb == CommandVerb.BG_DISP2 || c.Verb == CommandVerb.BG_DISPCG || c.Verb == CommandVerb.BG_FADE ||
-                    c.Verb == CommandVerb.BG_REVERT);
-                SKPaint palEffectPaint = PaletteEffectScriptParameter.IdentityPaint;
-                if (palCommand is not null && lastBgCommand is not null && commands.IndexOf(palCommand) > commands.IndexOf(lastBgCommand))
-                {
-                    switch (((PaletteEffectScriptParameter)palCommand.Parameters[0]).Effect)
-                    {
-                        case PaletteEffectScriptParameter.PaletteEffect.INVERTED:
-                            palEffectPaint = PaletteEffectScriptParameter.InvertedPaint;
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, bgSelectionButton));
                             break;
 
-                        case PaletteEffectScriptParameter.PaletteEffect.GRAYSCALE:
-                            palEffectPaint = PaletteEffectScriptParameter.GrayscalePaint;
+                        case ScriptParameter.ParameterType.BG_SCROLL_DIRECTION:
+                            ScriptCommandDropDown bgScrollDropDown = new() { Command = command, ParameterIndex = i };
+                            bgScrollDropDown.Items.AddRange(Enum.GetNames<BgScrollDirectionScriptParameter.BgScrollDirection>().Select(i => new ListItem { Text = i, Key = i }));
+                            bgScrollDropDown.SelectedKey = ((BgScrollDirectionScriptParameter)parameter).ScrollDirection.ToString();
+                            bgScrollDropDown.SelectedKeyChanged += BgScrollDropDown_SelectedKeyChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, bgScrollDropDown));
                             break;
 
-                        case PaletteEffectScriptParameter.PaletteEffect.SEPIA:
-                            palEffectPaint = PaletteEffectScriptParameter.SepiaPaint;
-                            break;
+                        case ScriptParameter.ParameterType.BGM:
+                            BgmScriptParameter bgmParam = (BgmScriptParameter)parameter;
+                            StackLayout bgmLink = ControlGenerator.GetFileLink(bgmParam.Bgm, _tabs, _log);
 
-                        case PaletteEffectScriptParameter.PaletteEffect.DIMMED:
-                            palEffectPaint = PaletteEffectScriptParameter.DimmedPaint;
-                            break;
-                    }
-                }
-                ScriptItemCommand bgScrollCommand = null;
-                for (int i = commands.Count - 1; i >= 0; i--)
-                {
-                    if (commands[i].Verb == CommandVerb.BG_REVERT)
-                    {
-                        bgReverted = true;
-                        continue;
-                    }
-                    if (commands[i].Verb == CommandVerb.BG_SCROLL)
-                    {
-                        bgScrollCommand = commands[i];
-                        continue;
-                    }
-                    // Checks to see if this is one of the commands that sets a BG_REVERT immune background or if BG_REVERT hasn't been called
-                    if (commands[i].Verb == CommandVerb.BG_DISP || commands[i].Verb == CommandVerb.BG_DISP2 ||
-                        (commands[i].Verb == CommandVerb.BG_FADE && (((BgScriptParameter)commands[i].Parameters[0]).Background is not null)) ||
-                        (!bgReverted && (commands[i].Verb == CommandVerb.BG_DISPCG || commands[i].Verb == CommandVerb.BG_FADE)))
-                    {
-                        BackgroundItem background = (commands[i].Verb == CommandVerb.BG_FADE && ((BgScriptParameter)commands[i].Parameters[0]).Background is null) ?
-                            ((BgScriptParameter)commands[i].Parameters[1]).Background : ((BgScriptParameter)commands[i].Parameters[0]).Background;
+                            ScriptCommandDropDown bgmDropDown = new() { Command = command, ParameterIndex = i, Link = (ClearableLinkButton)bgmLink.Items[1].Control };
+                            bgmDropDown.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.BGM).Select(i => new ListItem { Text = i.DisplayName, Key = i.DisplayName }));
+                            bgmDropDown.SelectedKey = bgmParam.Bgm.DisplayName;
+                            bgmDropDown.SelectedKeyChanged += BgmDropDown_SelectedKeyChanged;
 
-                        if (background is not null)
-                        {
-                            switch (background.BackgroundType)
+                            StackLayout bgmLayout = new()
                             {
-                                case BgType.TEX_CG_DUAL_SCREEN:
-                                    SKBitmap dualScreenBg = background.GetBackground();
-                                    if (bgScrollCommand is not null && ((BgScrollDirectionScriptParameter)bgScrollCommand.Parameters[0]).ScrollDirection == BgScrollDirectionScriptParameter.BgScrollDirection.DOWN)
-                                    {
-                                        canvas.DrawBitmap(dualScreenBg, new SKRect(0, background.Graphic2.Height - 192, 256, background.Graphic2.Height), new SKRect(0, 0, 256, 192));
-                                        int bottomScreenX = dualScreenBg.Height - 192;
-                                        canvas.DrawBitmap(dualScreenBg, new SKRect(0, bottomScreenX, 256, bottomScreenX + 192), new SKRect(0, 192, 256, 384));
-                                    }
-                                    else
-                                    {
-                                        canvas.DrawBitmap(dualScreenBg, new SKRect(0, 0, 256, 192), new SKRect(0, 0, 256, 192));
-                                        canvas.DrawBitmap(dualScreenBg, new SKRect(0, background.Graphic2.Height, 256, background.Graphic2.Height + 192), new SKRect(0, 192, 256, 384));
-                                    }
-                                    break;
-
-                                case BgType.TEX_CG_SINGLE:
-                                    if (((BoolScriptParameter)commands[i].Parameters[1]).Value
-                                        || (bgScrollCommand is not null && ((BgScrollDirectionScriptParameter)bgScrollCommand.Parameters[0]).ScrollDirection == BgScrollDirectionScriptParameter.BgScrollDirection.DOWN))
-                                    {
-                                        SKBitmap bgBitmap = background.GetBackground();
-                                        canvas.DrawBitmap(bgBitmap, new SKRect(0, bgBitmap.Height - 192, bgBitmap.Width, bgBitmap.Height),
-                                            new SKRect(0, 192, 256, 384));
-                                    }
-                                    else
-                                    {
-                                        canvas.DrawBitmap(background.GetBackground(), new SKPoint(0, 192));
-                                    }
-                                    break;
-
-                                case BgType.TEX_CG_WIDE:
-                                    if (bgScrollCommand is not null && ((BgScrollDirectionScriptParameter)bgScrollCommand.Parameters[0]).ScrollDirection == BgScrollDirectionScriptParameter.BgScrollDirection.RIGHT)
-                                    {
-                                        SKBitmap bgBitmap = background.GetBackground();
-                                        canvas.DrawBitmap(bgBitmap, new SKRect(bgBitmap.Width - 256, 0, bgBitmap.Width, 192), new SKRect(0, 192, 256, 384));
-                                    }
-                                    else
-                                    {
-                                        canvas.DrawBitmap(background.GetBackground(), new SKPoint(0, 192));
-                                    }
-                                    break;
-
-                                case BgType.TEX_CG:
-                                    canvas.DrawBitmap(background.GetBackground(), new SKPoint(0, 192));
-                                    break;
-
-                                default:
-                                    canvas.DrawBitmap(background.GetBackground(), new SKPoint(0, 192), palEffectPaint);
-                                    break;
+                                Orientation = Orientation.Horizontal,
+                                VerticalContentAlignment = VerticalAlignment.Center,
+                                Items =
+                            {
+                                ControlGenerator.GetControlWithLabel(parameter.Name, bgmDropDown),
+                                bgmLink,
                             }
-                            break;
-                        }
-                    }
-                }
-
-                // Draw character sprites
-                Dictionary<CharacterItem, PositionedSprite> sprites = new();
-                Dictionary<CharacterItem, PositionedSprite> previousSprites = new();
-
-                ScriptItemCommand previousCommand = null;
-                foreach (ScriptItemCommand command in commands)
-                {
-                    if (previousCommand?.Verb == CommandVerb.DIALOGUE)
-                    {
-                        SpriteExitScriptParameter spriteExitMoveParam = (SpriteExitScriptParameter)previousCommand?.Parameters[3]; // exits/moves happen _after_ dialogue is advanced, so we check these at this point
-                        if (spriteExitMoveParam.ExitTransition != SpriteExitScriptParameter.SpriteExitTransition.NO_EXIT)
-                        {
-                            CharacterItem prevCharacter = (CharacterItem)_project.Items.First(i => i.Type == ItemDescription.ItemType.Character &&
-                                i.Name == $"CHR_{_project.Characters[(int)((DialogueScriptParameter)previousCommand.Parameters[0]).Line.Speaker].Name}");
-                            SpriteScriptParameter previousSpriteParam = (SpriteScriptParameter)previousCommand.Parameters[1];
-                            short layer = ((ShortScriptParameter)previousCommand.Parameters[9]).Value;
-                            switch (spriteExitMoveParam.ExitTransition)
-                            {
-                                case SpriteExitScriptParameter.SpriteExitTransition.SLIDE_LEFT_TO_LEFT_FADE_OUT:
-                                case SpriteExitScriptParameter.SpriteExitTransition.SLIDE_LEFT_TO_RIGHT_FADE_OUT:
-                                case SpriteExitScriptParameter.SpriteExitTransition.SLIDE_FROM_CENTER_TO_LEFT_FADE_OUT:
-                                case SpriteExitScriptParameter.SpriteExitTransition.SLIDE_FROM_CENTER_TO_RIGHT_FADE_OUT:
-                                case SpriteExitScriptParameter.SpriteExitTransition.FADE_OUT_CENTER:
-                                case SpriteExitScriptParameter.SpriteExitTransition.FADE_OUT_LEFT:
-                                    if (sprites.ContainsKey(prevCharacter) && previousSprites.ContainsKey(prevCharacter) && ((SpriteScriptParameter)previousCommand.Parameters[1]).Sprite?.Sprite?.Character == prevCharacter.MessageInfo.Character)
-                                    {
-                                        sprites.Remove(prevCharacter);
-                                        previousSprites.Remove(prevCharacter);
-                                    }
-                                    break;
-
-                                case SpriteExitScriptParameter.SpriteExitTransition.SLIDE_CENTER_TO_LEFT_AND_STAY:
-                                case SpriteExitScriptParameter.SpriteExitTransition.SLIDE_RIGHT_TO_LEFT_AND_STAY:
-                                    sprites[prevCharacter] = new() { Sprite = previousSpriteParam.Sprite, Positioning = new() { Position = SpritePositioning.SpritePosition.LEFT, Layer = layer } };
-                                    break;
-
-                                case SpriteExitScriptParameter.SpriteExitTransition.SLIDE_CENTER_TO_RIGHT_AND_STAY:
-                                case SpriteExitScriptParameter.SpriteExitTransition.SLIDE_LEFT_TO_RIGHT_AND_STAY:
-                                    sprites[prevCharacter] = new() { Sprite = previousSpriteParam.Sprite, Positioning = new() { Position = SpritePositioning.SpritePosition.RIGHT, Layer = layer } };
-                                    break;
-                            }
-                        }
-                    }
-                    if (command.Verb == CommandVerb.DIALOGUE)
-                    {
-                        SpriteScriptParameter spriteParam = (SpriteScriptParameter)command.Parameters[1];
-                        SKPaint spritePaint = PaletteEffectScriptParameter.IdentityPaint;
-                        if (commands.IndexOf(palCommand) > commands.IndexOf(command))
-                        {
-                            spritePaint = ((PaletteEffectScriptParameter)palCommand.Parameters[0]).Effect switch
-                            {
-                                PaletteEffectScriptParameter.PaletteEffect.INVERTED => PaletteEffectScriptParameter.InvertedPaint,
-                                PaletteEffectScriptParameter.PaletteEffect.GRAYSCALE => PaletteEffectScriptParameter.GrayscalePaint,
-                                PaletteEffectScriptParameter.PaletteEffect.SEPIA => PaletteEffectScriptParameter.SepiaPaint,
-                                PaletteEffectScriptParameter.PaletteEffect.DIMMED => PaletteEffectScriptParameter.DimmedPaint,
-                                _ => PaletteEffectScriptParameter.IdentityPaint,
                             };
-                        }
-                        if (spriteParam.Sprite is not null)
-                        {
-                            CharacterItem character;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(bgmLayout);
+                            break;
+
+                        case ScriptParameter.ParameterType.BGM_MODE:
+                            ScriptCommandDropDown bgmModeDropDown = new() { Command = command, ParameterIndex = i };
+                            bgmModeDropDown.Items.AddRange(Enum.GetNames<BgmModeScriptParameter.BgmMode>().Select(i => new ListItem { Text = i, Key = i }));
+                            bgmModeDropDown.SelectedKey = ((BgmModeScriptParameter)parameter).Mode.ToString();
+                            bgmModeDropDown.SelectedKeyChanged += BgmModeDropDown_SelectedKeyChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, bgmModeDropDown));
+                            break;
+
+                        case ScriptParameter.ParameterType.BOOL:
+                            ScriptCommandCheckBox boolParameterCheckbox = new() { Command = command, ParameterIndex = i, Checked = ((BoolScriptParameter)parameter).Value };
+                            boolParameterCheckbox.CheckedChanged += BoolParameterCheckbox_CheckedChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, boolParameterCheckbox));
+                            break;
+
+                        case ScriptParameter.ParameterType.CHARACTER:
+                            ScriptCommandDropDown dialoguePropertyDropDown = new() { Command = command, ParameterIndex = i };
+                            dialoguePropertyDropDown.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Character)
+                                .Select(c => new ListItem { Key = c.DisplayName, Text = c.DisplayName[4..] }));
+                            dialoguePropertyDropDown.SelectedKey = ((DialoguePropertyScriptParameter)parameter).Character.Name;
+                            dialoguePropertyDropDown.SelectedKeyChanged += DialoguePropertyDropDown_SelectedKeyChanged;
+                            _currentSpeakerDropDown.OtherDropDowns.Add(dialoguePropertyDropDown);
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, dialoguePropertyDropDown));
+                            break;
+
+                        case ScriptParameter.ParameterType.CHESS_FILE:
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name,
+                                new TextBox { Text = ((ChessFileScriptParameter)parameter).ChessFileIndex.ToString() }));
+                            break;
+
+                        case ScriptParameter.ParameterType.CHESS_PIECE:
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name,
+                                new TextBox { Text = ((ChessPieceScriptParameter)parameter).ChessPiece.ToString() }));
+                            break;
+
+                        case ScriptParameter.ParameterType.CHESS_SPACE:
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name,
+                                new TextBox { Text = ((ChessSpaceScriptParameter)parameter).SpaceIndex.ToString() }));
+                            break;
+
+                        case ScriptParameter.ParameterType.CHIBI:
+                            ScriptCommandDropDown chibiDropDown = new() { Command = command, ParameterIndex = i };
+                            chibiDropDown.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Chibi).Select(i => new ListItem { Text = i.Name, Key = i.Name }));
+                            chibiDropDown.SelectedKey = ((ChibiScriptParameter)parameter).Chibi.Name;
+                            chibiDropDown.SelectedKeyChanged += ChibiDropDown_SelectedKeyChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, chibiDropDown));
+                            break;
+
+                        case ScriptParameter.ParameterType.CHIBI_EMOTE:
+                            ScriptCommandDropDown chibiEmoteDropDown = new() { Command = command, ParameterIndex = i };
+                            chibiEmoteDropDown.Items.AddRange(Enum.GetNames<ChibiEmoteScriptParameter.ChibiEmote>().Select(i => new ListItem { Text = i, Key = i }));
+                            chibiEmoteDropDown.SelectedKey = ((ChibiEmoteScriptParameter)parameter).Emote.ToString();
+                            chibiEmoteDropDown.SelectedKeyChanged += ChibiEmoteDropDown_SelectedKeyChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, chibiEmoteDropDown));
+                            break;
+
+                        case ScriptParameter.ParameterType.CHIBI_ENTER_EXIT:
+                            ScriptCommandDropDown chibiEnterExitDropDown = new() { Command = command, ParameterIndex = i };
+                            chibiEnterExitDropDown.Items.AddRange(Enum.GetNames<ChibiEnterExitScriptParameter.ChibiEnterExitType>().Select(i => new ListItem { Text = i, Key = i }));
+                            chibiEnterExitDropDown.SelectedKey = ((ChibiEnterExitScriptParameter)parameter).Mode.ToString();
+                            chibiEnterExitDropDown.SelectedKeyChanged += ChibiEnterExitDropDown_SelectedKeyChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, chibiEnterExitDropDown));
+                            break;
+
+                        case ScriptParameter.ParameterType.COLOR:
+                            ScriptCommandColorPicker colorPicker = new()
+                            {
+                                AllowAlpha = false,
+                                Value = ((ColorScriptParameter)parameter).Color.ToEtoDrawingColor(),
+                                Command = command,
+                                ParameterIndex = i,
+                            };
+                            colorPicker.ValueChanged += ColorPicker_ValueChanged;
+                            currentShort += 2;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, colorPicker));
+                            break;
+
+                        case ScriptParameter.ParameterType.CONDITIONAL:
+                            ScriptCommandTextBox conditionalBox = new() { Text = ((ConditionalScriptParameter)parameter).Conditional, Command = command, ParameterIndex = i };
+                            conditionalBox.TextChanged += ConditionalBox_TextChanged;
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name,
+                                conditionalBox));
+                            break;
+
+                        case ScriptParameter.ParameterType.COLOR_MONOCHROME:
+                            ScriptCommandDropDown colorMonochromeDropDown = new() { Command = command, ParameterIndex = i, CurrentShort = currentShort };
+                            colorMonochromeDropDown.Items.AddRange(Enum.GetNames<ColorMonochromeScriptParameter.ColorMonochrome>().Select(i => new ListItem { Text = i, Key = i }));
+                            colorMonochromeDropDown.SelectedKey = ((ColorMonochromeScriptParameter)parameter).ColorType.ToString();
+                            colorMonochromeDropDown.SelectedKeyChanged += ColorMonochromeDropDown_SelectedKeyChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, colorMonochromeDropDown));
+                            break;
+
+                        case ScriptParameter.ParameterType.DIALOGUE:
+                            DialogueScriptParameter dialogueParam = (DialogueScriptParameter)parameter;
+                            ScriptCommandDropDown speakerDropDown = new() { Command = command, ParameterIndex = i, OtherDropDowns = new() };
+                            speakerDropDown.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Character).Select(c => new ListItem { Key = c.DisplayName, Text = c.DisplayName[4..] }));
                             try
                             {
-                                character = (CharacterItem)_project.Items.First(i => i.Type == ItemDescription.ItemType.Character &&
-                                i.DisplayName == $"CHR_{_project.Characters[(int)((DialogueScriptParameter)command.Parameters[0]).Line.Speaker].Name}");
+                                speakerDropDown.SelectedKey = _project.Items.First(i => i.Type == ItemDescription.ItemType.Character && i.DisplayName == $"CHR_{_project.Characters[(int)dialogueParam.Line.Speaker].Name}").DisplayName;
                             }
                             catch (InvalidOperationException)
                             {
-                                _log.LogWarning($"Unable to determine speaking character in DIALOGUE command in {_script.DisplayName}.");
-                                using Stream noPreviewStream = Assembly.GetCallingAssembly().GetManifestResourceStream("SerialLoops.Graphics.ScriptPreviewError.png");
-                                canvas.DrawImage(SKImage.FromEncodedData(noPreviewStream), new SKPoint(0, 0));
-                                canvas.Flush();
-                                _preview.Items.Add(new SKGuiImage(previewBitmap));
+                                _log.LogError("Failed to find character item -- have you saved all of your changes to character names?");
                                 return;
                             }
-                            SpriteEntranceScriptParameter spriteEntranceParam = (SpriteEntranceScriptParameter)command.Parameters[2];
-                            short layer = ((ShortScriptParameter)command.Parameters[9]).Value;
 
-                            if (!sprites.ContainsKey(character) && spriteEntranceParam.EntranceTransition != SpriteEntranceScriptParameter.SpriteEntranceTransition.NO_TRANSITION)
+                            speakerDropDown.SelectedKeyChanged += SpeakerDropDown_SelectedKeyChanged;
+                            if (currentCol > 0)
                             {
-                                sprites.Add(character, new());
-                                previousSprites.Add(character, new());
+                                controlsTable.Rows.Add(new(new TableLayout { Spacing = new Size(5, 5) }));
+                                currentRow++;
+                                currentCol = 0;
                             }
-                            if (sprites.ContainsKey(character))
+                            _currentSpeakerDropDown = speakerDropDown;
+
+                            ScriptCommandTextArea dialogueTextArea = new()
                             {
-                                previousSprites[character] = sprites[character];
-                            }
-                            if (spriteEntranceParam.EntranceTransition != SpriteEntranceScriptParameter.SpriteEntranceTransition.NO_TRANSITION)
-                            {
-                                switch (spriteEntranceParam.EntranceTransition)
+                                Text = dialogueParam.Line.Text.GetSubstitutedString(_project),
+                                AcceptsReturn = true,
+                                Command = command,
+                                ParameterIndex = i,
+                            };
+                            dialogueTextArea.TextChanged += DialogueTextArea_TextChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabelTable(parameter.Name,
+                                new StackLayout
                                 {
-                                    case SpriteEntranceScriptParameter.SpriteEntranceTransition.FADE_TO_CENTER:
-                                    case SpriteEntranceScriptParameter.SpriteEntranceTransition.SLIDE_LEFT_TO_CENTER:
-                                    case SpriteEntranceScriptParameter.SpriteEntranceTransition.SLIDE_RIGHT_TO_CENTER:
-                                        sprites[character] = new() { Sprite = spriteParam.Sprite, Positioning = new() { Position = SpritePositioning.SpritePosition.CENTER, Layer = layer }, PalEffect = spritePaint };
+                                    Orientation = Orientation.Horizontal,
+                                    Items =
+                                    {
+                                    speakerDropDown,
+                                    new StackLayoutItem(dialogueTextArea, expand: true),
+                                    },
+                                }));
+                            controlsTable.Rows.Add(new(new TableLayout { Spacing = new Size(5, 5) }));
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows.Add(new());
+                            currentRow++;
+                            currentCol = 0;
+                            break;
+
+                        case ScriptParameter.ParameterType.EPISODE_HEADER:
+                            ScriptCommandDropDown epHeaderDropDown = new() { Command = command, ParameterIndex = i };
+                            epHeaderDropDown.Items.AddRange(Enum.GetNames<EpisodeHeaderScriptParameter.Episode>().Select(e => new ListItem { Key = e, Text = e }));
+                            epHeaderDropDown.SelectedKey = ((EpisodeHeaderScriptParameter)parameter).EpisodeHeaderIndex.ToString();
+                            epHeaderDropDown.SelectedKeyChanged += EpHeaderDropDown_SelectedKeyChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name,
+                                epHeaderDropDown));
+                            break;
+
+                        case ScriptParameter.ParameterType.FLAG:
+                            ScriptCommandTextBox flagTextBox = new() { Command = command, ParameterIndex = i, Text = ((FlagScriptParameter)parameter).FlagName };
+                            flagTextBox.TextChanged += FlagTextBox_TextChanged;
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, flagTextBox));
+                            break;
+
+                        case ScriptParameter.ParameterType.FRIENDSHIP_LEVEL:
+                            ScriptCommandDropDown friendshipLevelDropDown = new() { Command = command, ParameterIndex = i };
+                            friendshipLevelDropDown.Items.AddRange(Enum.GetNames<FriendshipLevelScriptParameter.FriendshipCharacter>().Select(f => new ListItem { Key = f, Text = f }));
+                            friendshipLevelDropDown.SelectedKey = ((FriendshipLevelScriptParameter)parameter).Character.ToString();
+                            friendshipLevelDropDown.SelectedIndexChanged += FriendshipLevelDropDown_SelectedIndexChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, friendshipLevelDropDown));
+                            break;
+
+                        case ScriptParameter.ParameterType.ITEM:
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name,
+                                new TextBox { Text = ((ItemScriptParameter)parameter).ItemIndex.ToString() }));
+                            break;
+
+                        case ScriptParameter.ParameterType.MAP:
+                            MapScriptParameter mapParam = (MapScriptParameter)parameter;
+                            ScriptCommandDropDown mapDropDown = new() { Command = command, ParameterIndex = i };
+                            mapDropDown.Items.Add(new ListItem { Text = "NONE", Key = "NONE" });
+                            mapDropDown.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Map).Select(i => new ListItem { Text = i.Name, Key = i.Name }));
+                            mapDropDown.SelectedKey = mapParam.Map?.Name ?? "NONE";
+                            mapDropDown.SelectedKeyChanged += MapDropDown_SelectedKeyChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, mapDropDown));
+                            break;
+
+                        case ScriptParameter.ParameterType.OPTION:
+                            OptionScriptParameter optionParam = (OptionScriptParameter)parameter;
+                            ScriptCommandDropDown optionDropDown = new() { Command = command, ParameterIndex = i };
+                            optionDropDown.Items.Add(new ListItem { Text = "NONE", Key = "0" });
+                            optionDropDown.Items.AddRange(_script.Event.ChoicesSection.Objects.Skip(1).SkipLast(1).Select(c => new ListItem { Text = c.Text.GetSubstitutedString(_project), Key = c.Id.ToString() }));
+                            optionDropDown.SelectedKey = optionParam.Option.Id.ToString();
+                            optionDropDown.SelectedKeyChanged += OptionDropDown_SelectedKeyChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, optionDropDown));
+                            break;
+
+                        case ScriptParameter.ParameterType.PALETTE_EFFECT:
+                            ScriptCommandDropDown paletteEffectDropDown = new() { Command = command, ParameterIndex = i };
+                            paletteEffectDropDown.Items.AddRange(Enum.GetNames<PaletteEffectScriptParameter.PaletteEffect>().Select(t => new ListItem { Text = t, Key = t }));
+                            paletteEffectDropDown.SelectedKey = ((PaletteEffectScriptParameter)parameter).Effect.ToString();
+                            paletteEffectDropDown.SelectedKeyChanged += PaletteEffectDropDown_SelectedKeyChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, paletteEffectDropDown));
+                            break;
+
+                        case ScriptParameter.ParameterType.PLACE:
+                            PlaceScriptParameter placeParam = (PlaceScriptParameter)parameter;
+                            CommandGraphicSelectionButton placeSelectionButton = new(placeParam.Place is not null ? placeParam.Place
+                                : NonePreviewableGraphic.PLACE, _tabs, _log)
+                            {
+                                Command = command,
+                                ParameterIndex = i,
+                                Project = _project,
+                            };
+                            placeSelectionButton.Items.Add(NonePreviewableGraphic.PLACE);
+                            placeSelectionButton.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Place)
+                                .Select(p => p as IPreviewableGraphic));
+                            placeSelectionButton.SelectedChanged.Executed += (obj, args) => PlaceSelectionButtonSelectedChanged_Executed(placeSelectionButton, args);
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, placeSelectionButton));
+                            break;
+
+                        case ScriptParameter.ParameterType.SCREEN:
+                            ScriptCommandScreenSelector screenSelector = new(_log, ((ScreenScriptParameter)parameter).Screen, true)
+                            {
+                                Command = command,
+                                ParameterIndex = i,
+                                CurrentShort = currentShort,
+                            };
+                            screenSelector.ScreenChanged += ScreenSelector_ScreenChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, screenSelector));
+                            break;
+
+                        case ScriptParameter.ParameterType.SCRIPT_SECTION:
+                            ScriptCommandDropDown scriptSectionDropDown = new() { Command = command, ParameterIndex = i, CurrentShort = i };
+                            if (command.Verb == CommandVerb.VGOTO)
+                            {
+                                scriptSectionDropDown.CurrentShort = 2;
+                            }
+                            scriptSectionDropDown.Items.AddRange(_script.Event.ScriptSections.Where(s => (_script.Event.LabelsSection.Objects.FirstOrDefault(l => l.Name.Replace("/", "") == s.Name)?.Id ?? 0) != 0).Select(s => new ListItem { Text = s.Name, Key = s.Name }));
+                            scriptSectionDropDown.SelectedKey = ((ScriptSectionScriptParameter)parameter)?.Section?.Name ?? "NONE";
+                            scriptSectionDropDown.SelectedKeyChanged += ScriptSectionDropDown_SelectedKeyChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, scriptSectionDropDown));
+                            break;
+
+                        case ScriptParameter.ParameterType.SFX:
+                            ScriptCommandNumericStepper sfxNumericStepper = new() { Command = command, ParameterIndex = i, Value = ((SfxScriptParameter)parameter).SfxIndex };
+                            sfxNumericStepper.ValueChanged += SfxNumericStepper_ValueChanged;
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, sfxNumericStepper));
+                            break;
+
+                        case ScriptParameter.ParameterType.SFX_MODE:
+                            ScriptCommandDropDown sfxModeDropDown = new() { Command = command, ParameterIndex = i };
+                            sfxModeDropDown.Items.AddRange(Enum.GetNames<SfxModeScriptParameter.SfxMode>().Select(t => new ListItem { Text = t, Key = t }));
+                            sfxModeDropDown.SelectedKey = ((SfxModeScriptParameter)parameter).Mode.ToString();
+                            sfxModeDropDown.SelectedKeyChanged += SfxModeDropDown_SelectedKeyChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, sfxModeDropDown));
+                            break;
+
+                        case ScriptParameter.ParameterType.SHORT:
+                            ScriptCommandNumericStepper shortNumericStepper = new()
+                            {
+                                Command = command,
+                                ParameterIndex = i,
+                                MinValue = short.MinValue,
+                                MaxValue = short.MaxValue,
+                                DecimalPlaces = 0,
+                                Value = ((ShortScriptParameter)parameter).Value
+                            };
+                            if (command.Verb == CommandVerb.SND_PLAY && parameter.Name == "Crossfade Time (Frames)")
+                            {
+                                shortNumericStepper.SecondIndex = 4;
+                            }
+                            if (command.Verb == CommandVerb.HARUHI_METER)
+                            {
+                                shortNumericStepper.ParameterIndex++;
+                            }
+                            shortNumericStepper.ValueChanged += ShortNumericStepper_ValueChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, shortNumericStepper));
+                            break;
+
+                        case ScriptParameter.ParameterType.SPRITE:
+                            SpriteScriptParameter spriteParam = (SpriteScriptParameter)parameter;
+                            CommandGraphicSelectionButton spriteSelectionButton = new(spriteParam.Sprite is not null ? spriteParam.Sprite
+                                : NonePreviewableGraphic.CHARACTER_SPRITE, _tabs, _log)
+                            {
+                                Command = command,
+                                ParameterIndex = i,
+                                Project = _project,
+                            };
+                            spriteSelectionButton.Items.Add(NonePreviewableGraphic.CHARACTER_SPRITE);
+                            spriteSelectionButton.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Character_Sprite).Select(s => (IPreviewableGraphic)s));
+                            spriteSelectionButton.SelectedChanged.Executed += (obj, args) => SpriteSelectionButton_SelectionMade(spriteSelectionButton, args);
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, spriteSelectionButton));
+                            break;
+
+                        case ScriptParameter.ParameterType.SPRITE_ENTRANCE:
+                            ScriptCommandDropDown spriteEntranceDropDown = new() { Command = command, ParameterIndex = i };
+                            spriteEntranceDropDown.Items.AddRange(Enum.GetNames<SpriteEntranceScriptParameter.SpriteEntranceTransition>().Select(t => new ListItem { Text = t, Key = t }));
+                            spriteEntranceDropDown.SelectedKey = ((SpriteEntranceScriptParameter)parameter).EntranceTransition.ToString();
+                            spriteEntranceDropDown.SelectedKeyChanged += SpriteEntranceDropDown_SelectedKeyChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, spriteEntranceDropDown));
+                            break;
+
+                        case ScriptParameter.ParameterType.SPRITE_EXIT:
+                            ScriptCommandDropDown spriteExitDropDown = new() { Command = command, ParameterIndex = i };
+                            spriteExitDropDown.Items.AddRange(Enum.GetNames<SpriteExitScriptParameter.SpriteExitTransition>().Select(t => new ListItem { Text = t, Key = t }));
+                            spriteExitDropDown.SelectedKey = ((SpriteExitScriptParameter)parameter).ExitTransition.ToString();
+                            spriteExitDropDown.SelectedKeyChanged += SpriteExitDropDown_SelectedKeyChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, spriteExitDropDown));
+                            break;
+
+                        case ScriptParameter.ParameterType.SPRITE_SHAKE:
+                            ScriptCommandDropDown spriteShakeDropDown = new() { Command = command, ParameterIndex = i };
+                            spriteShakeDropDown.Items.AddRange(Enum.GetNames<SpriteShakeScriptParameter.SpriteShakeEffect>().Select(t => new ListItem { Text = t, Key = t }));
+                            spriteShakeDropDown.SelectedKey = ((SpriteShakeScriptParameter)parameter).ShakeEffect.ToString();
+                            spriteShakeDropDown.SelectedKeyChanged += SpriteShakeDropDown_SelectedKeyChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, spriteShakeDropDown));
+                            break;
+
+                        case ScriptParameter.ParameterType.TEXT_ENTRANCE_EFFECT:
+                            ScriptCommandDropDown textEntranceEffectDropDown = new() { Command = command, ParameterIndex = i };
+                            textEntranceEffectDropDown.Items.AddRange(Enum.GetNames<TextEntranceEffectScriptParameter.TextEntranceEffect>().Select(t => new ListItem { Text = t, Key = t }));
+                            textEntranceEffectDropDown.SelectedKey = ((TextEntranceEffectScriptParameter)parameter).EntranceEffect.ToString();
+                            textEntranceEffectDropDown.SelectedKeyChanged += TextEntranceEffectDropDown_SelectedKeyChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, textEntranceEffectDropDown));
+                            break;
+
+                        case ScriptParameter.ParameterType.TOPIC:
+                            string topicName = _project.Items.FirstOrDefault(i => i.Type == ItemDescription.ItemType.Topic &&
+                                ((TopicItem)i).Topic.Id == ((TopicScriptParameter)parameter).TopicId)?.DisplayName;
+                            if (string.IsNullOrEmpty(topicName))
+                            {
+                                // If the topic has been deleted, we will just display the index in a textbox
+                                TopicSelectButton setUpTopicControlButton = new() { Text = "Select a Topic", ScriptCommand = command, ParameterIndex = i };
+
+                                StackLayout deletedTopicLayout = new()
+                                {
+                                    Orientation = Orientation.Horizontal,
+                                    Items =
+                                {
+                                    new TextBox { Text = ((TopicScriptParameter)parameter).TopicId.ToString() },
+                                    setUpTopicControlButton,
+                                },
+                                };
+
+                                setUpTopicControlButton.Layout = deletedTopicLayout;
+                                setUpTopicControlButton.Click += SetUpTopicControlButton_Click;
+
+                                ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                    ControlGenerator.GetControlWithLabel(parameter.Name,
+                                    deletedTopicLayout));
+                            }
+                            else
+                            {
+                                StackLayout topicLink = ControlGenerator.GetFileLink(_project.Items.FirstOrDefault(i => i.Type == ItemDescription.ItemType.Topic &&
+                                    ((TopicItem)i).Topic.Id == ((TopicScriptParameter)parameter).TopicId), _tabs, _log);
+
+                                ScriptCommandDropDown topicDropDown = new() { Command = command, ParameterIndex = i, Link = (ClearableLinkButton)topicLink.Items[1].Control };
+                                topicDropDown.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Topic)
+                                    .Select(t => new ListItem { Key = t.DisplayName, Text = t.DisplayName }));
+                                topicDropDown.SelectedKey = topicName;
+                                topicDropDown.SelectedIndexChanged += TopicDropDown_SelectedIndexChanged;
+
+                                StackLayout topicLinkLayout = new()
+                                {
+                                    Orientation = Orientation.Horizontal,
+                                    Items =
+                                {
+                                    topicDropDown,
+                                    topicLink,
+                                },
+                                };
+
+                                ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                    ControlGenerator.GetControlWithLabel(parameter.Name,
+                                    topicLinkLayout));
+                            }
+                            break;
+
+                        case ScriptParameter.ParameterType.TRANSITION:
+                            ScriptCommandDropDown transitionDropDown = new() { Command = command, ParameterIndex = i };
+                            transitionDropDown.Items.AddRange(Enum.GetNames<TransitionScriptParameter.TransitionEffect>().Select(t => new ListItem { Text = t, Key = t }));
+                            transitionDropDown.SelectedKey = ((TransitionScriptParameter)parameter).Transition.ToString();
+                            transitionDropDown.SelectedKeyChanged += TransitionDropDown_SelectedKeyChanged;
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(
+                                ControlGenerator.GetControlWithLabel(parameter.Name, transitionDropDown));
+                            break;
+
+                        case ScriptParameter.ParameterType.VOICE_LINE:
+                            VoicedLineScriptParameter vceParam = (VoicedLineScriptParameter)parameter;
+                            StackLayout vceLink = ControlGenerator.GetFileLink(vceParam.VoiceLine is not null ? vceParam.VoiceLine : NoneItem.VOICE, _tabs, _log);
+
+                            ScriptCommandDropDown vceDropDown = new() { Command = command, ParameterIndex = i, Link = (ClearableLinkButton)vceLink.Items[1].Control };
+                            vceDropDown.Items.Add(new ListItem { Key = "NONE", Text = "NONE" });
+                            vceDropDown.Items.AddRange(_project.Items.Where(i => i.Type == ItemDescription.ItemType.Voice).Select(i => new ListItem { Text = i.Name, Key = i.Name }));
+                            vceDropDown.SelectedKey = vceParam.VoiceLine?.Name ?? "NONE";
+                            vceDropDown.SelectedKeyChanged += VceDropDown_SelectedKeyChanged;
+
+                            StackLayout vceLayout = new()
+                            {
+                                Orientation = Orientation.Horizontal,
+                                VerticalContentAlignment = VerticalAlignment.Center,
+                                Items =
+                            {
+                                ControlGenerator.GetControlWithLabel(parameter.Name, vceDropDown),
+                                vceLink,
+                            }
+                            };
+
+                            ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows[0].Cells.Add(vceLayout);
+                            break;
+
+                        default:
+                            _log.LogError($"Invalid parameter detected in script {_script.Name} parameter {parameter.Name}");
+                            break;
+                    }
+                    currentCol++;
+                    if (currentCol >= cols)
+                    {
+                        currentCol = 0;
+                        currentRow++;
+                        controlsTable.Rows.Add(new(new TableLayout { Spacing = new Size(5, 5) }));
+                        ((TableLayout)controlsTable.Rows.Last().Cells[0].Control).Rows.Add(new());
+                    }
+                    currentShort++;
+                }
+
+                _editorControls.Items.Add(new StackLayoutItem(controlsTable, expand: true));
+            }
+            catch (Exception ex)
+            {
+                _log.LogException("Failed to load editor controls!", ex);
+            }
+        }
+
+        private void UpdatePreview()
+        {
+            try
+            {
+                _preview.Items.Clear();
+
+                SKBitmap previewBitmap = new(256, 384);
+                SKCanvas canvas = new(previewBitmap);
+                canvas.DrawColor(SKColors.Black);
+
+                ScriptItemCommand currentCommand = ((ScriptCommandSectionEntry)_commandsPanel.Viewer.SelectedItem)?.Command;
+                if (currentCommand is not null)
+                {
+                    List<ScriptItemCommand> commands = currentCommand.WalkCommandGraph(_commands, _script.Graph);
+
+                    if (commands is null)
+                    {
+                        _log.LogWarning($"No path found to current command.");
+                        using Stream noPreviewStream = Assembly.GetCallingAssembly().GetManifestResourceStream("SerialLoops.Graphics.ScriptPreviewError.png");
+                        canvas.DrawImage(SKImage.FromEncodedData(noPreviewStream), new SKPoint(0, 0));
+                        canvas.Flush();
+                        _preview.Items.Add(new SKGuiImage(previewBitmap));
+                        return;
+                    }
+
+                    if (commands.Any(c => c.Verb == CommandVerb.EPHEADER)
+                        && ((EpisodeHeaderScriptParameter)commands.Last(c => c.Verb == CommandVerb.EPHEADER).Parameters[0])
+                            .EpisodeHeaderIndex != EpisodeHeaderScriptParameter.Episode.None)
+                    {
+                        // We use names here because display names can change but names cannot
+                        SystemTextureItem systexItem = ((EpisodeHeaderScriptParameter)commands.Last(c => c.Verb == CommandVerb.EPHEADER).Parameters[0]).EpisodeHeaderIndex switch
+                        {
+                            EpisodeHeaderScriptParameter.Episode.EPISODE_1 =>
+                                (SystemTextureItem)_project.Items.First(i => i.Name == "SYSTEX_SYS_CMN_T60"),
+                            EpisodeHeaderScriptParameter.Episode.EPISODE_2 =>
+                                (SystemTextureItem)_project.Items.First(i => i.Name == "SYSTEX_SYS_CMN_T61"),
+                            EpisodeHeaderScriptParameter.Episode.EPISODE_3 =>
+                                (SystemTextureItem)_project.Items.First(i => i.Name == "SYSTEX_SYS_CMN_T62"),
+                            EpisodeHeaderScriptParameter.Episode.EPISODE_4 =>
+                                (SystemTextureItem)_project.Items.First(i => i.Name == "SYSTEX_SYS_CMN_T63"),
+                            EpisodeHeaderScriptParameter.Episode.EPISODE_5 =>
+                                (SystemTextureItem)_project.Items.First(i => i.Name == "SYSTEX_SYS_CMN_T64"),
+                            EpisodeHeaderScriptParameter.Episode.EPILOGUE =>
+                                (SystemTextureItem)_project.Items.First(i => i.Name == "SYSTEX_SYS_CMN_T66"),
+                            _ => null,
+                        };
+
+                        canvas.DrawBitmap(systexItem.GetTexture(), new SKPoint(0, 0));
+                    }
+                    else
+                    {
+                        // Draw top screen "kinetic" background
+                        for (int i = commands.Count - 1; i >= 0; i--)
+                        {
+                            if (commands[i].Verb == CommandVerb.KBG_DISP && ((BgScriptParameter)commands[i].Parameters[0]).Background is not null)
+                            {
+                                canvas.DrawBitmap(((BgScriptParameter)commands[i].Parameters[0]).Background.GetBackground(), new SKPoint(0, 0));
+                                break;
+                            }
+                        }
+
+                        // Draw Place
+                        for (int i = commands.Count - 1; i >= 0; i--)
+                        {
+                            if (commands[i].Verb == CommandVerb.SET_PLACE)
+                            {
+                                if (((BoolScriptParameter)commands[i].Parameters[0]).Value && (((PlaceScriptParameter)commands[i].Parameters[1]).Place is not null))
+                                {
+                                    canvas.DrawBitmap(((PlaceScriptParameter)commands[i].Parameters[1]).Place.GetPreview(_project), new SKPoint(5, 40));
+                                }
+                                break;
+                            }
+                        }
+
+                        // Draw top screen chibis
+                        List<ChibiItem> chibis = new();
+
+                        foreach (StartingChibiEntry chibi in _script.Event.StartingChibisSection?.Objects ?? new List<StartingChibiEntry>())
+                        {
+                            if (chibi.ChibiIndex > 0)
+                            {
+                                chibis.Add((ChibiItem)_project.Items.First(i => i.Type == ItemDescription.ItemType.Chibi && ((ChibiItem)i).ChibiIndex == chibi.ChibiIndex));
+                            }
+                        }
+                        for (int i = 0; i < commands.Count; i++)
+                        {
+                            if (commands[i].Verb == CommandVerb.CHIBI_ENTEREXIT)
+                            {
+                                if (((ChibiEnterExitScriptParameter)commands[i].Parameters[1]).Mode == ChibiEnterExitScriptParameter.ChibiEnterExitType.ENTER)
+                                {
+                                    ChibiItem chibi = ((ChibiScriptParameter)commands[i].Parameters[0]).Chibi;
+                                    if (!chibis.Contains(chibi))
+                                    {
+                                        if (chibi.ChibiIndex < 1 || chibis.Count == 0)
+                                        {
+                                            chibis.Add(chibi);
+                                        }
+                                        else
+                                        {
+                                            bool inserted = false;
+                                            for (int j = 0; j < chibis.Count; j++)
+                                            {
+                                                if (chibis[j].ChibiIndex > chibi.ChibiIndex)
+                                                {
+                                                    chibis.Insert(j, chibi);
+                                                    inserted = true;
+                                                    break;
+                                                }
+                                            }
+                                            if (!inserted)
+                                            {
+                                                chibis.Add(chibi);
+                                            }
+                                        }
+                                    }
+                                    else
+                                    {
+                                        _log.LogWarning($"Chibi {chibi.Name} set to join, but already was present");
+                                    }
+                                }
+                                else
+                                {
+                                    try
+                                    {
+                                        chibis.Remove(((ChibiScriptParameter)commands[i].Parameters[0]).Chibi);
+                                    }
+                                    catch (Exception)
+                                    {
+                                        _log.LogWarning($"Chibi set to leave was not present.");
+                                    }
+                                }
+                            }
+                        }
+
+                        int chibiStartX, chibiY;
+                        if (commands.Any(c => c.Verb == EventFile.CommandVerb.OP_MODE))
+                        {
+                            chibiStartX = 100;
+                            chibiY = 50;
+                        }
+                        else
+                        {
+                            chibiStartX = 44;
+                            chibiY = 100;
+                        }
+                        int chibiCurrentX = chibiStartX;
+                        int chibiWidth = 0;
+                        foreach (ChibiItem chibi in chibis)
+                        {
+                            SKBitmap chibiFrame = chibi.ChibiAnimations.First().Value.ElementAt(0).Frame;
+                            canvas.DrawBitmap(chibiFrame, new SKPoint(chibiCurrentX, chibiY));
+                            chibiWidth = chibiFrame.Width - 10;
+                            chibiCurrentX += chibiWidth;
+                        }
+
+                        // Draw top screen chibi emotes
+                        if (currentCommand.Verb == CommandVerb.CHIBI_EMOTE)
+                        {
+                            ChibiItem chibi = ((ChibiScriptParameter)currentCommand.Parameters[0]).Chibi;
+                            if (chibis.Contains(chibi))
+                            {
+                                int chibiIndex = chibis.IndexOf(chibi);
+                                SKBitmap emotes = _project.Grp.Files.First(f => f.Name == "SYS_ADV_T08DNX").GetImage(width: 32, transparentIndex: 0);
+                                int internalYOffset = ((int)((ChibiEmoteScriptParameter)currentCommand.Parameters[1]).Emote - 1) * 32;
+                                int externalXOffset = chibiStartX + chibiWidth * chibiIndex;
+                                canvas.DrawBitmap(emotes, new SKRect(0, internalYOffset, 32, internalYOffset + 32), new SKRect(externalXOffset + 16, chibiY - 32, externalXOffset + 48, chibiY));
+                            }
+                            else
+                            {
+                                _log.LogWarning($"Chibi {chibi.Name} not currently on screen; cannot display emote.");
+                            }
+                        }
+                    }
+
+                    // Draw background
+                    bool bgReverted = false;
+                    ScriptItemCommand palCommand = commands.LastOrDefault(c => c.Verb == CommandVerb.PALEFFECT);
+                    ScriptItemCommand lastBgCommand = commands.LastOrDefault(c => c.Verb == CommandVerb.BG_DISP ||
+                        c.Verb == CommandVerb.BG_DISP2 || c.Verb == CommandVerb.BG_DISPCG || c.Verb == CommandVerb.BG_FADE ||
+                        c.Verb == CommandVerb.BG_REVERT);
+                    SKPaint palEffectPaint = PaletteEffectScriptParameter.IdentityPaint;
+                    if (palCommand is not null && lastBgCommand is not null && commands.IndexOf(palCommand) > commands.IndexOf(lastBgCommand))
+                    {
+                        switch (((PaletteEffectScriptParameter)palCommand.Parameters[0]).Effect)
+                        {
+                            case PaletteEffectScriptParameter.PaletteEffect.INVERTED:
+                                palEffectPaint = PaletteEffectScriptParameter.InvertedPaint;
+                                break;
+
+                            case PaletteEffectScriptParameter.PaletteEffect.GRAYSCALE:
+                                palEffectPaint = PaletteEffectScriptParameter.GrayscalePaint;
+                                break;
+
+                            case PaletteEffectScriptParameter.PaletteEffect.SEPIA:
+                                palEffectPaint = PaletteEffectScriptParameter.SepiaPaint;
+                                break;
+
+                            case PaletteEffectScriptParameter.PaletteEffect.DIMMED:
+                                palEffectPaint = PaletteEffectScriptParameter.DimmedPaint;
+                                break;
+                        }
+                    }
+                    ScriptItemCommand bgScrollCommand = null;
+                    for (int i = commands.Count - 1; i >= 0; i--)
+                    {
+                        if (commands[i].Verb == CommandVerb.BG_REVERT)
+                        {
+                            bgReverted = true;
+                            continue;
+                        }
+                        if (commands[i].Verb == CommandVerb.BG_SCROLL)
+                        {
+                            bgScrollCommand = commands[i];
+                            continue;
+                        }
+                        // Checks to see if this is one of the commands that sets a BG_REVERT immune background or if BG_REVERT hasn't been called
+                        if (commands[i].Verb == CommandVerb.BG_DISP || commands[i].Verb == CommandVerb.BG_DISP2 ||
+                            (commands[i].Verb == CommandVerb.BG_FADE && (((BgScriptParameter)commands[i].Parameters[0]).Background is not null)) ||
+                            (!bgReverted && (commands[i].Verb == CommandVerb.BG_DISPCG || commands[i].Verb == CommandVerb.BG_FADE)))
+                        {
+                            BackgroundItem background = (commands[i].Verb == CommandVerb.BG_FADE && ((BgScriptParameter)commands[i].Parameters[0]).Background is null) ?
+                                ((BgScriptParameter)commands[i].Parameters[1]).Background : ((BgScriptParameter)commands[i].Parameters[0]).Background;
+
+                            if (background is not null)
+                            {
+                                switch (background.BackgroundType)
+                                {
+                                    case BgType.TEX_CG_DUAL_SCREEN:
+                                        SKBitmap dualScreenBg = background.GetBackground();
+                                        if (bgScrollCommand is not null && ((BgScrollDirectionScriptParameter)bgScrollCommand.Parameters[0]).ScrollDirection == BgScrollDirectionScriptParameter.BgScrollDirection.DOWN)
+                                        {
+                                            canvas.DrawBitmap(dualScreenBg, new SKRect(0, background.Graphic2.Height - 192, 256, background.Graphic2.Height), new SKRect(0, 0, 256, 192));
+                                            int bottomScreenX = dualScreenBg.Height - 192;
+                                            canvas.DrawBitmap(dualScreenBg, new SKRect(0, bottomScreenX, 256, bottomScreenX + 192), new SKRect(0, 192, 256, 384));
+                                        }
+                                        else
+                                        {
+                                            canvas.DrawBitmap(dualScreenBg, new SKRect(0, 0, 256, 192), new SKRect(0, 0, 256, 192));
+                                            canvas.DrawBitmap(dualScreenBg, new SKRect(0, background.Graphic2.Height, 256, background.Graphic2.Height + 192), new SKRect(0, 192, 256, 384));
+                                        }
                                         break;
 
-                                    case SpriteEntranceScriptParameter.SpriteEntranceTransition.FADE_IN_LEFT:
-                                    case SpriteEntranceScriptParameter.SpriteEntranceTransition.PEEK_RIGHT_TO_LEFT:
-                                    case SpriteEntranceScriptParameter.SpriteEntranceTransition.SLIDE_RIGHT_TO_LEFT:
-                                    case SpriteEntranceScriptParameter.SpriteEntranceTransition.SLIDE_RIGHT_TO_LEFT_FAST:
-                                    case SpriteEntranceScriptParameter.SpriteEntranceTransition.SLIDE_RIGHT_TO_LEFT_SLOW:
-                                        sprites[character] = new() { Sprite = spriteParam.Sprite, Positioning = new() { Position = SpritePositioning.SpritePosition.LEFT, Layer = layer }, PalEffect = spritePaint };
+                                    case BgType.TEX_CG_SINGLE:
+                                        if (((BoolScriptParameter)commands[i].Parameters[1]).Value
+                                            || (bgScrollCommand is not null && ((BgScrollDirectionScriptParameter)bgScrollCommand.Parameters[0]).ScrollDirection == BgScrollDirectionScriptParameter.BgScrollDirection.DOWN))
+                                        {
+                                            SKBitmap bgBitmap = background.GetBackground();
+                                            canvas.DrawBitmap(bgBitmap, new SKRect(0, bgBitmap.Height - 192, bgBitmap.Width, bgBitmap.Height),
+                                                new SKRect(0, 192, 256, 384));
+                                        }
+                                        else
+                                        {
+                                            canvas.DrawBitmap(background.GetBackground(), new SKPoint(0, 192));
+                                        }
                                         break;
 
-                                    case SpriteEntranceScriptParameter.SpriteEntranceTransition.SLIDE_LEFT_TO_RIGHT:
-                                    case SpriteEntranceScriptParameter.SpriteEntranceTransition.SLIDE_LEFT_TO_RIGHT_FAST:
-                                    case SpriteEntranceScriptParameter.SpriteEntranceTransition.SLIDE_LEFT_TO_RIGHT_SLOW:
-                                        sprites[character] = new() { Sprite = spriteParam.Sprite, Positioning = new() { Position = SpritePositioning.SpritePosition.RIGHT, Layer = layer }, PalEffect = spritePaint };
+                                    case BgType.TEX_CG_WIDE:
+                                        if (bgScrollCommand is not null && ((BgScrollDirectionScriptParameter)bgScrollCommand.Parameters[0]).ScrollDirection == BgScrollDirectionScriptParameter.BgScrollDirection.RIGHT)
+                                        {
+                                            SKBitmap bgBitmap = background.GetBackground();
+                                            canvas.DrawBitmap(bgBitmap, new SKRect(bgBitmap.Width - 256, 0, bgBitmap.Width, 192), new SKRect(0, 192, 256, 384));
+                                        }
+                                        else
+                                        {
+                                            canvas.DrawBitmap(background.GetBackground(), new SKPoint(0, 192));
+                                        }
+                                        break;
+
+                                    case BgType.TEX_CG:
+                                        canvas.DrawBitmap(background.GetBackground(), new SKPoint(0, 192));
+                                        break;
+
+                                    default:
+                                        canvas.DrawBitmap(background.GetBackground(), new SKPoint(0, 192), palEffectPaint);
+                                        break;
+                                }
+                                break;
+                            }
+                        }
+                    }
+
+                    // Draw character sprites
+                    Dictionary<CharacterItem, PositionedSprite> sprites = new();
+                    Dictionary<CharacterItem, PositionedSprite> previousSprites = new();
+
+                    ScriptItemCommand previousCommand = null;
+                    foreach (ScriptItemCommand command in commands)
+                    {
+                        if (previousCommand?.Verb == CommandVerb.DIALOGUE)
+                        {
+                            SpriteExitScriptParameter spriteExitMoveParam = (SpriteExitScriptParameter)previousCommand?.Parameters[3]; // exits/moves happen _after_ dialogue is advanced, so we check these at this point
+                            if (spriteExitMoveParam.ExitTransition != SpriteExitScriptParameter.SpriteExitTransition.NO_EXIT)
+                            {
+                                CharacterItem prevCharacter = (CharacterItem)_project.Items.First(i => i.Type == ItemDescription.ItemType.Character &&
+                                    i.Name == $"CHR_{_project.Characters[(int)((DialogueScriptParameter)previousCommand.Parameters[0]).Line.Speaker].Name}");
+                                SpriteScriptParameter previousSpriteParam = (SpriteScriptParameter)previousCommand.Parameters[1];
+                                short layer = ((ShortScriptParameter)previousCommand.Parameters[9]).Value;
+                                switch (spriteExitMoveParam.ExitTransition)
+                                {
+                                    case SpriteExitScriptParameter.SpriteExitTransition.SLIDE_LEFT_TO_LEFT_FADE_OUT:
+                                    case SpriteExitScriptParameter.SpriteExitTransition.SLIDE_LEFT_TO_RIGHT_FADE_OUT:
+                                    case SpriteExitScriptParameter.SpriteExitTransition.SLIDE_FROM_CENTER_TO_LEFT_FADE_OUT:
+                                    case SpriteExitScriptParameter.SpriteExitTransition.SLIDE_FROM_CENTER_TO_RIGHT_FADE_OUT:
+                                    case SpriteExitScriptParameter.SpriteExitTransition.FADE_OUT_CENTER:
+                                    case SpriteExitScriptParameter.SpriteExitTransition.FADE_OUT_LEFT:
+                                        if (sprites.ContainsKey(prevCharacter) && previousSprites.ContainsKey(prevCharacter) && ((SpriteScriptParameter)previousCommand.Parameters[1]).Sprite?.Sprite?.Character == prevCharacter.MessageInfo.Character)
+                                        {
+                                            sprites.Remove(prevCharacter);
+                                            previousSprites.Remove(prevCharacter);
+                                        }
+                                        break;
+
+                                    case SpriteExitScriptParameter.SpriteExitTransition.SLIDE_CENTER_TO_LEFT_AND_STAY:
+                                    case SpriteExitScriptParameter.SpriteExitTransition.SLIDE_RIGHT_TO_LEFT_AND_STAY:
+                                        sprites[prevCharacter] = new() { Sprite = previousSpriteParam.Sprite, Positioning = new() { Position = SpritePositioning.SpritePosition.LEFT, Layer = layer } };
+                                        break;
+
+                                    case SpriteExitScriptParameter.SpriteExitTransition.SLIDE_CENTER_TO_RIGHT_AND_STAY:
+                                    case SpriteExitScriptParameter.SpriteExitTransition.SLIDE_LEFT_TO_RIGHT_AND_STAY:
+                                        sprites[prevCharacter] = new() { Sprite = previousSpriteParam.Sprite, Positioning = new() { Position = SpritePositioning.SpritePosition.RIGHT, Layer = layer } };
                                         break;
                                 }
                             }
-                            else if (sprites.ContainsKey(character))
+                        }
+                        if (command.Verb == CommandVerb.DIALOGUE)
+                        {
+                            SpriteScriptParameter spriteParam = (SpriteScriptParameter)command.Parameters[1];
+                            SKPaint spritePaint = PaletteEffectScriptParameter.IdentityPaint;
+                            if (commands.IndexOf(palCommand) > commands.IndexOf(command))
                             {
-                                sprites[character] = new() { Sprite = spriteParam.Sprite, Positioning = sprites[character].Positioning, PalEffect = spritePaint };
+                                spritePaint = ((PaletteEffectScriptParameter)palCommand.Parameters[0]).Effect switch
+                                {
+                                    PaletteEffectScriptParameter.PaletteEffect.INVERTED => PaletteEffectScriptParameter.InvertedPaint,
+                                    PaletteEffectScriptParameter.PaletteEffect.GRAYSCALE => PaletteEffectScriptParameter.GrayscalePaint,
+                                    PaletteEffectScriptParameter.PaletteEffect.SEPIA => PaletteEffectScriptParameter.SepiaPaint,
+                                    PaletteEffectScriptParameter.PaletteEffect.DIMMED => PaletteEffectScriptParameter.DimmedPaint,
+                                    _ => PaletteEffectScriptParameter.IdentityPaint,
+                                };
+                            }
+                            if (spriteParam.Sprite is not null)
+                            {
+                                CharacterItem character;
+                                try
+                                {
+                                    character = (CharacterItem)_project.Items.First(i => i.Type == ItemDescription.ItemType.Character &&
+                                    i.DisplayName == $"CHR_{_project.Characters[(int)((DialogueScriptParameter)command.Parameters[0]).Line.Speaker].Name}");
+                                }
+                                catch (InvalidOperationException)
+                                {
+                                    _log.LogWarning($"Unable to determine speaking character in DIALOGUE command in {_script.DisplayName}.");
+                                    using Stream noPreviewStream = Assembly.GetCallingAssembly().GetManifestResourceStream("SerialLoops.Graphics.ScriptPreviewError.png");
+                                    canvas.DrawImage(SKImage.FromEncodedData(noPreviewStream), new SKPoint(0, 0));
+                                    canvas.Flush();
+                                    _preview.Items.Add(new SKGuiImage(previewBitmap));
+                                    return;
+                                }
+                                SpriteEntranceScriptParameter spriteEntranceParam = (SpriteEntranceScriptParameter)command.Parameters[2];
+                                short layer = ((ShortScriptParameter)command.Parameters[9]).Value;
+
+                                if (!sprites.ContainsKey(character) && spriteEntranceParam.EntranceTransition != SpriteEntranceScriptParameter.SpriteEntranceTransition.NO_TRANSITION)
+                                {
+                                    sprites.Add(character, new());
+                                    previousSprites.Add(character, new());
+                                }
+                                if (sprites.ContainsKey(character))
+                                {
+                                    previousSprites[character] = sprites[character];
+                                }
+                                if (spriteEntranceParam.EntranceTransition != SpriteEntranceScriptParameter.SpriteEntranceTransition.NO_TRANSITION)
+                                {
+                                    switch (spriteEntranceParam.EntranceTransition)
+                                    {
+                                        case SpriteEntranceScriptParameter.SpriteEntranceTransition.FADE_TO_CENTER:
+                                        case SpriteEntranceScriptParameter.SpriteEntranceTransition.SLIDE_LEFT_TO_CENTER:
+                                        case SpriteEntranceScriptParameter.SpriteEntranceTransition.SLIDE_RIGHT_TO_CENTER:
+                                            sprites[character] = new() { Sprite = spriteParam.Sprite, Positioning = new() { Position = SpritePositioning.SpritePosition.CENTER, Layer = layer }, PalEffect = spritePaint };
+                                            break;
+
+                                        case SpriteEntranceScriptParameter.SpriteEntranceTransition.FADE_IN_LEFT:
+                                        case SpriteEntranceScriptParameter.SpriteEntranceTransition.PEEK_RIGHT_TO_LEFT:
+                                        case SpriteEntranceScriptParameter.SpriteEntranceTransition.SLIDE_RIGHT_TO_LEFT:
+                                        case SpriteEntranceScriptParameter.SpriteEntranceTransition.SLIDE_RIGHT_TO_LEFT_FAST:
+                                        case SpriteEntranceScriptParameter.SpriteEntranceTransition.SLIDE_RIGHT_TO_LEFT_SLOW:
+                                            sprites[character] = new() { Sprite = spriteParam.Sprite, Positioning = new() { Position = SpritePositioning.SpritePosition.LEFT, Layer = layer }, PalEffect = spritePaint };
+                                            break;
+
+                                        case SpriteEntranceScriptParameter.SpriteEntranceTransition.SLIDE_LEFT_TO_RIGHT:
+                                        case SpriteEntranceScriptParameter.SpriteEntranceTransition.SLIDE_LEFT_TO_RIGHT_FAST:
+                                        case SpriteEntranceScriptParameter.SpriteEntranceTransition.SLIDE_LEFT_TO_RIGHT_SLOW:
+                                            sprites[character] = new() { Sprite = spriteParam.Sprite, Positioning = new() { Position = SpritePositioning.SpritePosition.RIGHT, Layer = layer }, PalEffect = spritePaint };
+                                            break;
+                                    }
+                                }
+                                else if (sprites.ContainsKey(character))
+                                {
+                                    sprites[character] = new() { Sprite = spriteParam.Sprite, Positioning = sprites[character].Positioning, PalEffect = spritePaint };
+                                }
                             }
                         }
-                    }
-                    else if (command.Verb == CommandVerb.INVEST_START)
-                    {
-                        sprites.Clear();
-                        previousSprites.Clear();
-                    }
-                    previousCommand = command;
-                }
-
-                foreach (PositionedSprite sprite in sprites.Values.OrderBy(p => p.Positioning.Layer))
-                {
-                    SKBitmap spriteBitmap = sprite.Sprite.GetClosedMouthAnimation(_project)[0].Frame;
-                    canvas.DrawBitmap(spriteBitmap, sprite.Positioning.GetSpritePosition(spriteBitmap), sprite.PalEffect);
-                }
-
-                // Draw dialogue
-                ScriptItemCommand lastDialogueCommand = commands.LastOrDefault(c => c.Verb == CommandVerb.DIALOGUE);
-                if (commands.FindLastIndex(c => c.Verb == CommandVerb.TOGGLE_DIALOGUE &&
-                    !((BoolScriptParameter)c.Parameters[0]).Value) < commands.IndexOf(lastDialogueCommand))
-                {
-                    DialogueLine line = ((DialogueScriptParameter)lastDialogueCommand.Parameters[0]).Line;
-                    SKPaint dialoguePaint = line.Speaker switch
-                    {
-                        Speaker.MONOLOGUE => DialogueScriptParameter.Paint01,
-                        Speaker.INFO => DialogueScriptParameter.Paint04,
-                        _ => DialogueScriptParameter.Paint00,
-                    };
-                    if (!string.IsNullOrEmpty(line.Text))
-                    {
-                        canvas.DrawBitmap(_project.DialogueBitmap, new SKRect(0, 24, 32, 36), new SKRect(0, 344, 256, 356));
-                        SKColor dialogueBoxColor = _project.DialogueBitmap.GetPixel(0, 28);
-                        canvas.DrawRect(0, 356, 224, 384, new() { Color = dialogueBoxColor });
-                        canvas.DrawBitmap(_project.DialogueBitmap, new SKRect(0, 37, 32, 64), new SKRect(224, 356, 256, 384));
-                        canvas.DrawBitmap(_project.SpeakerBitmap, new SKRect(0, 16 * ((int)line.Speaker - 1), 64, 16 * ((int)line.Speaker)),
-                            new SKRect(0, 332, 64, 348));
-
-                        DrawText(line.Text, canvas, dialoguePaint, _project);
-                    }
-                }
-            }
-
-            canvas.Flush();
-
-            _preview.Items.Add(new SKGuiImage(previewBitmap));
-        }
-
-        public static void DrawText(string text, SKCanvas canvas, SKPaint color, Project project, int x = 10, int y = 352, bool formatting = true)
-        {
-            int currentX = x;
-            int currentY = y;
-
-            for (int i = 0; i < text.Length; i++)
-            {
-                // handle newlines
-                if (text[i] == '\n')
-                {
-                    currentX = 10;
-                    currentY += 14;
-                    continue;
-                }
-
-                // handle operators
-                if (formatting)
-                {
-                    if (i < text.Length - 2 && Regex.IsMatch(text[i..(i + 2)], @"\$\d"))
-                    {
-                        if (i < text.Length - 3 && Regex.IsMatch(text[i..(i + 3)], @"\$\d{2}"))
+                        else if (command.Verb == CommandVerb.INVEST_START)
                         {
-                            i++;
+                            sprites.Clear();
+                            previousSprites.Clear();
                         }
-                        i++;
-                        continue;
+                        previousCommand = command;
                     }
-                    else if (i < text.Length - 3 && Regex.IsMatch(text[i..(i + 3)], @"#W\d"))
+
+                    foreach (PositionedSprite sprite in sprites.Values.OrderBy(p => p.Positioning.Layer))
                     {
-                        if (i < text.Length - 4 && Regex.IsMatch(text[i..(i + 4)], @"#W\d{2}"))
-                        {
-                            i++;
-                        }
-                        i += 2;
-                        continue;
+                        SKBitmap spriteBitmap = sprite.Sprite.GetClosedMouthAnimation(_project)[0].Frame;
+                        canvas.DrawBitmap(spriteBitmap, sprite.Positioning.GetSpritePosition(spriteBitmap), sprite.PalEffect);
                     }
-                    else if (i < text.Length - 4 && Regex.IsMatch(text[i..(i + 4)], @"#P\d{2}"))
+
+                    // Draw dialogue
+                    ScriptItemCommand lastDialogueCommand = commands.LastOrDefault(c => c.Verb == CommandVerb.DIALOGUE);
+                    if (commands.FindLastIndex(c => c.Verb == CommandVerb.TOGGLE_DIALOGUE &&
+                        !((BoolScriptParameter)c.Parameters[0]).Value) < commands.IndexOf(lastDialogueCommand))
                     {
-                        color = int.Parse(Regex.Match(text[i..(i + 4)], @"#P(?<id>\d{2})").Groups["id"].Value) switch
+                        DialogueLine line = ((DialogueScriptParameter)lastDialogueCommand.Parameters[0]).Line;
+                        SKPaint dialoguePaint = line.Speaker switch
                         {
-                            1 => DialogueScriptParameter.Paint01,
-                            2 => DialogueScriptParameter.Paint02,
-                            3 => DialogueScriptParameter.Paint03,
-                            4 => DialogueScriptParameter.Paint04,
-                            5 => DialogueScriptParameter.Paint05,
-                            6 => DialogueScriptParameter.Paint06,
-                            7 => DialogueScriptParameter.Paint07,
+                            Speaker.MONOLOGUE => DialogueScriptParameter.Paint01,
+                            Speaker.INFO => DialogueScriptParameter.Paint04,
                             _ => DialogueScriptParameter.Paint00,
                         };
-                        i += 3;
-                        continue;
-                    }
-                    else if (i < text.Length - 3 && text[i..(i + 3)] == "#DP")
-                    {
-                        i += 3;
-                    }
-                    else if (i < text.Length - 6 && Regex.IsMatch(text[i..(i + 6)], @"#SE\d{3}"))
-                    {
-                        i += 6;
-                    }
-                    else if (i < text.Length - 4 && text[i..(i + 4)] == "#SK0")
-                    {
-                        i += 4;
-                    }
-                    else if (i < text.Length - 3 && text[i..(i + 3)] == "#sk")
-                    {
-                        i += 3;
+                        if (!string.IsNullOrEmpty(line.Text))
+                        {
+                            canvas.DrawBitmap(_project.DialogueBitmap, new SKRect(0, 24, 32, 36), new SKRect(0, 344, 256, 356));
+                            SKColor dialogueBoxColor = _project.DialogueBitmap.GetPixel(0, 28);
+                            canvas.DrawRect(0, 356, 224, 384, new() { Color = dialogueBoxColor });
+                            canvas.DrawBitmap(_project.DialogueBitmap, new SKRect(0, 37, 32, 64), new SKRect(224, 356, 256, 384));
+                            canvas.DrawBitmap(_project.SpeakerBitmap, new SKRect(0, 16 * ((int)line.Speaker - 1), 64, 16 * ((int)line.Speaker)),
+                                new SKRect(0, 332, 64, 348));
+
+                            Shared.DrawText(line.Text, canvas, dialoguePaint, _project);
+                        }
                     }
                 }
 
-                if (text[i] != '　') // if it's a space, we just skip drawing
-                {
-                    int charIndex = project.FontMap.CharMap.IndexOf(text[i]);
-                    if ((charIndex + 1) * 16 <= project.FontBitmap.Height)
-                    {
-                        canvas.DrawBitmap(project.FontBitmap, new SKRect(0, charIndex * 16, 16, (charIndex + 1) * 16),
-                            new SKRect(currentX, currentY, currentX + 16, currentY + 16), color);
-                    }
-                }
-                FontReplacement replacement = project.FontReplacement.ReverseLookup(text[i]);
-                if (replacement is not null && project.LangCode != "ja")
-                {
-                    currentX += replacement.Offset;
-                }
-                else
-                {
-                    currentX += 14;
-                }
+                canvas.Flush();
+
+                _preview.Items.Add(new SKGuiImage(previewBitmap));
+            }
+            catch (Exception ex)
+            {
+                _log.LogException("Failed to update preview!", ex);
             }
         }
