+++ src/SerialLoops/Editors/VoicedLineEditor.cs
@@ -220,7 +220,7 @@ namespace SerialLoops.Editors
             {
                 for (int i = 0; i <= 1; i++)
                 {
-                    ScriptEditor.DrawText(
+                    Shared.DrawText(
                         _subtitle,
                         canvas,
                         DialogueScriptParameter.Paint07,
@@ -232,7 +232,7 @@ namespace SerialLoops.Editors
                 }
             }
 
-            ScriptEditor.DrawText(
+            Shared.DrawText(
                 _subtitle,
                 canvas,
                 DialogueScriptParameter.Paint00,
